% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/each.R
\name{accross}
\alias{accross}
\title{Apply a set of functions to a set of columns}
\usage{
accross(select, funs = identity, .unpack_to = "{fn}_{var}",
  .fn_in_vars = FALSE)
}
\arguments{
\item{select}{tidy selection of columns, forwarded to \code{\link[=pick]{pick()}}}

\item{funs}{Functions to apply to each of the selected columns. Possible
values are:
\itemize{
\item A single function
\item A single quosure style lambda, e.g. \code{~ mean(.x, na.rm = TRUE)}
\item A named list of functions and/or lambdas
}

@return A tibble}
}
\description{
Creates a data frame by applying a set of functions to a tidy
selection of columns in the current slice
}
