% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tally.r
\name{add_tally}
\alias{add_count}
\alias{add_count_}
\alias{add_tally}
\alias{add_tally_}
\title{Add a column counting or tallying observations within groups}
\usage{
add_tally(x, wt, sort = FALSE)

add_tally_(x, wt = NULL, sort = FALSE)

add_count(x, ..., wt = NULL, sort = FALSE)

add_count_(x, vars, wt = NULL, sort = FALSE)
}
\arguments{
\item{x}{a \code{tbl}.}

\item{wt}{(Optional) If omitted, will count the number of rows. Otherwise,
performs a weighted tally by summing the (non-missing) values of variable wt}

\item{sort}{Whether to sort the result in descending order of n}

\item{..., vars}{Variables to group by.}
}
\description{
\code{add_tally} adds a column "n" to a table based on the number
of items within each existing group, while \code{add_count} is a shortcut that
does the grouping as well. These functions are to \code{\link{tally}}
and \code{\link{count}} as \code{\link{mutate}} is to \code{\link{summarise}}:
they add an additional column rather than collapsing each group.
}
\details{
\code{add_count} counts within the current groups of the data when present,
and preserves those groups (it does not add the new ones).
}
\examples{

add_tally(mtcars)
add_tally(group_by(mtcars, cyl))
add_tally(group_by(mtcars, cyl), sort = TRUE)

add_count(mtcars, cyl)
add_count(mtcars, cyl, am)

if (require("Lahman")) {
batting_tbl <- tbl_df(Batting)

# get records of players who played in multiple stints in the same year
batting_tbl \%>\%
  add_count(playerID, yearID) \%>\%
  filter(n > 1)

# get only players who played in more than three stints total
batting_tbl \%>\%
  add_count(playerID) \%>\%
  filter(n > 3)

# get only players with at least 1000 ABs
batting_tbl \%>\%
  add_count(playerID, wt = AB) \%>\%
  filter(n >= 1000)
}

}

