% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{args_by}
\alias{args_by}
\title{Helper for consistent documentation of \code{.by}}
\arguments{
\item{.by}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

<\code{\link[=dplyr_tidy_select]{tidy-select}}> Optionally, select columns to group
by. This grouping will only be active for the duration of this verb.

Can't be used when the data is already a grouped or rowwise data frame.}
}
\description{
Use \verb{@inheritParams args_by} to consistently document \code{.by}.
}
\keyword{internal}
