\name{as.data.table.source_dt}
\alias{as.data.table.source_dt}
\title{Coerce data table to source.}
\usage{
  as.data.table.source_dt(x, row.names = NULL,
    optional = NULL, ...)

  as.data.table.source_dt(x, row.names = NULL,
    optional = NULL, ...)
}
\description{
  Coerce data table to source.
}
\keyword{internal}

