\name{as.source}
\alias{as.source}
\title{Convert an existing object to a data source.}
\usage{
  as.source(x, ...)
}
\arguments{
  \item{x}{an object to convert}

  \item{...}{other arguments passed to methods}
}
\description{
  Convert an existing object to a data source.
}
\examples{
as.source(mtcars)
}

