% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-cube.r
\name{as.table.tbl_cube}
\alias{as.table.tbl_cube}
\alias{as.data.frame.tbl_cube}
\alias{as_tibble.tbl_cube}
\title{Coerce a \code{tbl_cube} to other data structures}
\usage{
\method{as.table}{tbl_cube}(x, ..., measure = 1L)

\method{as.data.frame}{tbl_cube}(x, ...)

\method{as_tibble}{tbl_cube}(x, ...)
}
\arguments{
\item{x}{a \code{tbl_cube}}

\item{...}{Passed on to individual methods; otherwise ignored.}

\item{measure}{A measure name or index, default: the first measure}
}
\description{
Supports conversion to tables, data frames, tibbles.

For a cube, the data frame returned by
\code{\link[tibble:as_tibble]{tibble::as_tibble()}} resulting data frame contains the
dimensions as character values (and not as factors).
}
