% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R, R/tbl-cube.r
\name{as_data_frame}
\alias{as_data_frame}
\alias{as_data_frame.grouped_df}
\alias{as_data_frame.tbl_cube}
\title{Convert to a data frame}
\usage{
\method{as_data_frame}{grouped_df}(x, ...)

\method{as_data_frame}{tbl_cube}(x, ...)
}
\arguments{
\item{x}{A list. Each element of the list must have the same length.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Functions that convert the input to a \code{data_frame}.
}
\details{
For a grouped data frame, the \code{\link[tibble]{as_data_frame}}
S3 generic simply removes the grouping.

For a cube, the resulting data frame contains the dimensions
  as character values.
}
\seealso{
\code{\link[tibble]{as_data_frame}}
}

