% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-s3.r
\name{backend_src}
\alias{backend_src}
\alias{backend_src}
\alias{src_desc}
\alias{backend_src}
\alias{sql_translate_env}
\alias{backend_src}
\alias{sql_translate_env.NULL}
\title{Source generics.}
\usage{
src_desc(x)

sql_translate_env(con)

\method{sql_translate_env}{NULL}(con)
}
\description{
These generics retrieve metadata for a given src.
}
\keyword{internal}
