% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bands.R
\docType{data}
\name{band}
\alias{band}
\alias{instrument}
\alias{instrument2}
\title{Band membership}
\format{A tibble}
\usage{
band
}
\description{
These toy data sets describe band members of the Beatles and Rolling Stones.
They are designed to provide concise examples of dplyr joins that use data
that can be included in its entirety on a lecture slide.
}
\examples{
left_join(band, instrument, by = "name")
right_join(band, instrument, by = "name")
full_join(band, instrument, by = "name")
inner_join(band, instrument, by = "name")
semi_join(band, instrument, by = "name")
anti_join(band, instrument, by = "name")
left_join(band, instrument2, by = c("name" = "artist"))
}
