% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.r
\name{bunch_by}
\alias{bunch_by}
\title{Alternative to \code{\link[=group_by]{group_by()}} using vctrs}
\usage{
bunch_by(.data, ..., .drop = group_by_drop_default(.data))
}
\arguments{
\item{.data}{tibble}

\item{...}{selection of columns to group by}
}
\description{
Alternative to \code{\link[=group_by]{group_by()}} using vctrs
}
