% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{columns}
\alias{columns}
\title{Select columns}
\usage{
columns(...)
}
\arguments{
\item{...}{Variables to include/exclude in mutate/summarise. You
can use same specifications as in \code{\link{select}}. If
missing, defaults to all non-grouping variables.}
}
\description{
This helper has equivalent semantics to \code{\link{select}()}. Its
purpose is to provide \code{select()} semantics to the colwise
summarising and mutating verbs.
}
\seealso{
\code{\link{summarise_all}()}
}

