% Generated by roxygen2 (4.0.0): do not edit by hand
\name{colwise}
\alias{colwise}
\alias{colwise_q}
\title{colwise (standard evaluation).}
\usage{
colwise(tbl, calls, ...)

colwise_q(tbl, calls, vars)
}
\arguments{
\item{tbl}{Input data.}

\item{calls}{List of function calls, generated by \code{\link{funs}}}

\item{vars,...}{List of quoted objects specifying variables to select.}
}
\description{
Apply one or more functions to one or more columns.
}
\examples{
# One function
colwise(iris, funs(mean))
colwise(iris, funs(mean), Petal.Width)
colwise(iris, funs(mean), -Species)
colwise(iris, funs(mean), matches("Sepal"))

# Two functions
colwise(iris, funs(min, max))
colwise(iris, funs(min, max), Petal.Width)
colwise(iris, funs(min, max), matches("Sepal"))

# Alternative function specification
colwise(iris, funs(max(., na.rm = TRUE))

fs <- c("min", "max")
colwise(iris, funs_q(fs))

# Alternative variable specification
vars <- names(iris)
colwise_q(iris, funs(max), vars)
}

