% Generated by roxygen2 (4.0.0): do not edit by hand
\name{colwise_q}
\alias{colwise_q}
\title{colwise (standard evaluation).}
\usage{
colwise_q(tbl, calls, vars = dots())
}
\arguments{
\item{tbl}{Input data.}

\item{calls}{List of function calls, generated by \code{\link{funs}}}

\item{vars}{List of quoted objects specifying variables to select.}
}
\description{
Apply one or more functions to one or more columns.
}
\examples{
# One function
colwise_q(iris, funs("mean"))
colwise_q(iris, funs("mean"), dots(Petal.Width))
colwise_q(iris, funs("mean"), dots(-Species))
colwise_q(iris, funs("mean"), dots(matches("Sepal")))

# Two functions
colwise_q(iris, funs("min", "max"))
colwise_q(iris, funs("min", "max"), dots(Petal.Width))
colwise_q(iris, funs("min", "max"), dots(matches("Sepal")))

# Alternative function specification
colwise_q(iris, funs(max, max(., na.rm = TRUE)))
colwise_q(iris, funs("mean"), dots(x))
}

