% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumall}
\alias{cumall}
\alias{cumany}
\alias{cummean}
\title{Cumulativate versions of any, all, and mean}
\usage{
cumall(x)

cumany(x)

cummean(x)
}
\arguments{
\item{x}{For \code{cumall()} and \code{cumany()}, a logical vector; for
\code{cummean()} an integer or numeric vector}
}
\description{
dplyr adds \code{cumall()}, \code{cumany()}, and \code{cummean()} to complete
R's set of cumulate functions to match the aggregation functions available
in most databases
}
\examples{
x <- c(1, 3, 5, 2, 2)

# `cummean()` returns a numeric/integer vector of the same length
# as the input vector.
cummean(x)

# `cumall()` and `cumany()` return logicals, not numerics/integers.
cumall(x)
cumany(x)


# An NA only affects all elements after it for `cumall()` and `cummean()`,
# but NOT for `cumany()`
x <- c(1, 3, 5, NA, 2, 2)

# NAs in every position after the corresponding NA position in the vector.
cumall(x)
cummean(x)

# No NAs at all!
cumany(x)

# An NA at the beginning of the vector causes all elements after to be NA.
x <- c(NA, 1, 3, 5, 2, 2)

cummean(x)
cumall(x)
cumany(x)

# NULLs are thrown out.
x <- c(1, 3, 5, NULL, 2, 2)

cummean(x)
cumall(x)
cumany(x)
}
