% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/each.R
\name{current_key}
\alias{current_key}
\title{key data frame for the current group}
\usage{
current_key()
}
\value{
A data frame, with a single row, and one column per grouping variable
}
\description{
key data frame for the current group
}
