% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbi-database}
\alias{db_data_type}
\alias{db_has_table}
\alias{db_list_tables}
\alias{dbi-database}
\title{Database generics.}
\usage{
db_list_tables(con)

db_has_table(con, table)

db_data_type(con, fields)
}
\description{
These three generics are used to get information about the database.
dplyr provides default methods for DBIConnection which call the DBI
equivalents, \code{dbListTables}, \code{dbExistsTable} and \code{dbDataType}.
It should be rarely necessary to provide more specific methods for
DBI compliant interfaces.
}
\keyword{internal}

