% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbi-sql}
\alias{dbi-sql}
\alias{sql_analyze}
\alias{sql_begin}
\alias{sql_commit}
\alias{sql_create_index}
\alias{sql_create_table}
\alias{sql_drop_table}
\alias{sql_explain}
\alias{sql_insert_into}
\alias{sql_rollback}
\alias{sql_select}
\alias{sql_subquery}
\title{SQL generics.}
\usage{
sql_begin(con, ...)

sql_commit(con, ...)

sql_rollback(con, ...)

sql_create_table(con, table, types, temporary = FALSE, ...)

sql_insert_into(con, table, values, ...)

sql_create_index(con, table, columns, name = NULL, ...)

sql_drop_table(con, table, force = FALSE, ...)

sql_analyze(con, table, ...)

sql_select(con, select, from, where = NULL, group_by = NULL,
  having = NULL, order_by = NULL, limit = NULL, offset = NULL, ...)

sql_subquery(con, sql, name = random_table_name(), ...)

sql_explain(con, sql, ...)
}
\arguments{
\item{con}{A database connection.}
}
\description{
These generics are used to run various types of SQL queries.
Default methods are provides for \code{DBIConnectdion}, but variations in
SQL across databases means that it's likely that each backend will require
a few variations of these.
}
\section{copy_to}{

Currently, the only user of \code{sql_begin()}, \code{sql_commit()},
\code{sql_rollback()}, \code{sql_create_table()}, \code{sql_insert_into()},
\code{sql_create_indexes()}, \code{sql_drop_table()} and
\code{sql_analyze()}. If you find yourself overriding many of these
functions it may suggest that you should just override \code{\link{copy_to}}
instead.
}
\keyword{internal}

