% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-each.R
\name{defunct-each}
\alias{defunct-each}
\alias{summarise_each}
\alias{summarise_each_}
\alias{mutate_each}
\alias{mutate_each_}
\alias{summarize_each}
\alias{summarize_each_}
\title{Defunct functions for working with multiple columns}
\usage{
summarise_each(tbl, funs, ...)

summarise_each_(tbl, funs, vars)

mutate_each(tbl, funs, ...)

mutate_each_(tbl, funs, vars)

summarize_each(tbl, funs, ...)

summarize_each_(tbl, funs, vars)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{mutate_each()} and \code{summarise_each()} are deprecated in favour of
the new \code{\link[=across]{across()}} function that works within \code{summarise()} and \code{mutate()}.
}
\keyword{internal}
