% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-context.R
\name{deprec-context}
\alias{deprec-context}
\alias{cur_data}
\alias{cur_data_all}
\title{Information about the "current" group or variable}
\usage{
cur_data()

cur_data_all()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions were deprecated in dplyr 1.1.0 in favor of \code{\link[=pick]{pick()}}.
\itemize{
\item \code{cur_data()} gives the current data for the current group (excluding
grouping variables).
\item \code{cur_data_all()} gives the current data for the current group (including
grouping variables).
}
}
\keyword{internal}
