% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-format.r
\name{dim_desc}
\alias{dim_desc}
\title{Describing dimensions}
\usage{
dim_desc(x)
}
\arguments{
\item{x}{Object to show dimensions for.}
}
\description{
Prints the dimensions of an array-like object in a user-friendly manner,
substituting \code{NA} with ?? (for SQL queries).
}
\examples{
dim_desc(mtcars)
}
\keyword{internal}
