% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{package}
\name{dplyr-package}
\alias{dplyr}
\alias{dplyr-package}
\title{dplyr: A Grammar of Data Manipulation}
\description{
To learn more about dplyr, start with the vignettes:
\code{browseVignettes(package = "dplyr")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dplyr.tidyverse.org}
  \item \url{https://github.com/tidyverse/dplyr}
  \item Report bugs at \url{https://github.com/tidyverse/dplyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})

Authors:
\itemize{
  \item Romain François (\href{https://orcid.org/0000-0002-2444-4226}{ORCID})
  \item Lionel Henry
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
  \item Davis Vaughan \email{davis@posit.co} (\href{https://orcid.org/0000-0003-4777-038X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
