% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dplyr.r
\docType{package}
\name{dplyr}
\alias{dplyr}
\alias{dplyr-package}
\title{dplyr: a grammar of data manipulation}
\description{
dplyr provides a flexible grammar of data manipulation. It's the next
iteration of plyr, focussed on tools for working with data frames (hence the
\emph{d} in the name).
}
\details{
It has three main goals:

\itemize{
\item Identify the most important data manipulation verbs and make them
  easy to use from R.
\item Provide blazing fast performance for in-memory data by writing key
  pieces in C++ (using Rcpp)
\item Use the same interface to work with data no matter where it's stored,
  whether in a data frame, a data table or database.
}

To learn more about dplyr, start with the vignettes:
\code{browseVignettes(package = "dplyr")}
}

