% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{dplyr_by}
\alias{dplyr_by}
\title{Grouping with \code{.by}}
\description{
There are two ways to group in dplyr:
\itemize{
\item Persistent grouping with \code{\link[=group_by]{group_by()}}
\item Temporary grouping with \code{.by}
}

This help page is dedicated to explaining where and why you might want to use the latter.
The most important reason to use \code{.by} is that you \emph{never} have to remember to use \code{\link[=ungroup]{ungroup()}} after it.
The grouping is always temporary, meaning that if an ungrouped data frame goes in, then an ungrouped data frame comes out, regardless of the number of grouping columns that are specified.

The following dplyr verbs support \code{.by}:
\itemize{
\item \code{\link[=mutate]{mutate()}}
\item \code{\link[=summarise]{summarise()}}
\item \code{\link[=filter]{filter()}}
\item \code{\link[=slice]{slice()}} and its variants, such as \code{\link[=slice_head]{slice_head()}}
}

Let's take a look at the two grouping approaches using this \code{expenses} data set, which tracks costs accumulated across various \code{id}s and \code{region}s:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses <- tibble(
  id = c(1, 2, 1, 3, 1, 2, 3),
  region = c("A", "A", "A", "B", "B", "A", "A"),
  cost = c(25, 20, 19, 12, 9, 6, 6)
)
expenses
#> # A tibble: 7 x 3
#>      id region  cost
#>   <dbl> <chr>  <dbl>
#> 1     1 A         25
#> 2     2 A         20
#> 3     1 A         19
#> 4     3 B         12
#> 5     1 B          9
#> 6     2 A          6
#> 7     3 A          6
}\if{html}{\out{</div>}}

Imagine that you wanted to compute the average cost per region.
You'd probably write something like this:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  group_by(region) \%>\%
  summarise(cost = mean(cost))
#> # A tibble: 2 x 2
#>   region  cost
#>   <chr>  <dbl>
#> 1 A       15.2
#> 2 B       10.5
}\if{html}{\out{</div>}}

As of dplyr 1.1.0, an additional option is available that lets you specify the grouping \emph{inline} within the verb:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  summarise(cost = mean(cost), .by = region)
#> # A tibble: 2 x 2
#>   region  cost
#>   <chr>  <dbl>
#> 1 A       15.2
#> 2 B       10.5
}\if{html}{\out{</div>}}

This great idea comes from \href{https://CRAN.R-project.org/package=data.table}{data.table}, where the equivalent syntax looks something like \code{expenses[, .(cost = mean(cost)), by = region]}.

Grouping with \code{.by} is temporary, meaning that since \code{expenses} was an ungrouped data frame, the result after applying \code{.by} will also always be an ungrouped data frame, full stop.
Compare that with \code{group_by() \%>\% summarise()}, where \code{summarise()} generally peels off 1 layer of grouping by default, with a message that it is doing so if there were originally more than one grouping columns:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  group_by(id, region) \%>\%
  summarise(cost = mean(cost))
#> `summarise()` has grouped output by 'id'. You can override using the `.groups`
#> argument.
#> # A tibble: 5 x 3
#> # Groups:   id [3]
#>      id region  cost
#>   <dbl> <chr>  <dbl>
#> 1     1 A         22
#> 2     1 B          9
#> 3     2 A         13
#> 4     3 A          6
#> 5     3 B         12
}\if{html}{\out{</div>}}

This behavior is sometimes useful to sequentially "roll up" a data frame that has been grouped by multiple columns, but often you just want to group temporarily.
Traditionally, you'd either \code{ungroup()} after the \code{summarise()} or explicitly set \code{.groups = "drop"} to achieve this:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  group_by(id, region) \%>\%
  summarise(cost = mean(cost), .groups = "drop")
#> # A tibble: 5 x 3
#>      id region  cost
#>   <dbl> <chr>  <dbl>
#> 1     1 A         22
#> 2     1 B          9
#> 3     2 A         13
#> 4     3 A          6
#> 5     3 B         12
}\if{html}{\out{</div>}}

Because \code{.by} grouping is temporary, you don't need to worry about ungrouping and it never needs to emit a message to remind you what is happening:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  summarise(cost = mean(cost), .by = c(id, region))
#> # A tibble: 5 x 3
#>      id region  cost
#>   <dbl> <chr>  <dbl>
#> 1     1 A         22
#> 2     1 B          9
#> 3     2 A         13
#> 4     3 A          6
#> 5     3 B         12
}\if{html}{\out{</div>}}

Note that we specified multiple columns to group by using the \link[=dplyr_tidy_select]{tidy-select} syntax \code{c(id, region)}.
If you have a character vector of column names you'd like to group by, you can do so with \code{.by = all_of(my_cols)}.

To prevent surprising results, you can't use \code{.by} on an existing grouped data frame:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\% 
  group_by(id) \%>\%
  summarise(cost = mean(cost), .by = c(id, region))
#> Error in `summarise()`:
#> ! `.by` can't be supplied when `.data` is a grouped data frame.
}\if{html}{\out{</div>}}

So far we've focused on the usage of \code{.by} with \code{summarise()}, but \code{.by} works with a number of other dplyr verbs.
For example, you could append the mean cost per region onto the original data frame as a new column rather than computing a summary:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  mutate(cost_by_region = mean(cost), .by = region)
#> # A tibble: 7 x 4
#>      id region  cost cost_by_region
#>   <dbl> <chr>  <dbl>          <dbl>
#> 1     1 A         25           15.2
#> 2     2 A         20           15.2
#> 3     1 A         19           15.2
#> 4     3 B         12           10.5
#> 5     1 B          9           10.5
#> 6     2 A          6           15.2
#> 7     3 A          6           15.2
}\if{html}{\out{</div>}}

Or you could slice out the maximum cost per combination of id and region:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  slice_max(cost, n = 1, by = c(id, region))
#> # A tibble: 5 x 3
#>      id region  cost
#>   <dbl> <chr>  <dbl>
#> 1     1 A         25
#> 2     1 B          9
#> 3     2 A         20
#> 4     3 A          6
#> 5     3 B         12
}\if{html}{\out{</div>}}

Again, note that the result of this \code{slice_max()} is an ungrouped data frame, which is probably what you'd expect here.
Compare that to the \code{group_by()} approach, which uses persistent grouping and returns another grouped data frame:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{expenses \%>\%
  group_by(id, region) \%>\%
  slice_max(cost, n = 1)
#> # A tibble: 5 x 3
#> # Groups:   id, region [5]
#>      id region  cost
#>   <dbl> <chr>  <dbl>
#> 1     1 A         25
#> 2     1 B          9
#> 3     2 A         20
#> 4     3 A          6
#> 5     3 B         12
}\if{html}{\out{</div>}}
\subsection{Verbs without \code{.by} support}{

If a dplyr verb doesn't support \code{.by}, then that typically means that the verb isn't inherently affected by grouping.
For example, \code{\link[=pull]{pull()}} and \code{\link[=rename]{rename()}} don't support \code{.by}, because specifying columns to group by would not affect their implementations.

That said, there are a few exceptions to this where sometimes a dplyr verb doesn't support \code{.by}, but \emph{does} have special support for grouped data frames created by \code{\link[=group_by]{group_by()}}.
This is typically because the verbs are required to retain the grouping columns, for example:
\itemize{
\item \code{\link[=select]{select()}} always retains grouping columns, with a message if any aren't specified in the \code{select()} call.
\item \code{\link[=distinct]{distinct()}} and \code{\link[=count]{count()}} place unspecified grouping columns at the front of the data frame before computing their results.
\item \code{\link[=arrange]{arrange()}} has a \code{.by_group} argument to optionally order by grouping columns first.
}

If \code{group_by()} didn't exist, then these verbs would not have special support for grouped data frames.
}
}
\keyword{internal}
