% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{dplyr_tidy_eval}
\alias{dplyr_tidy_eval}
\title{Argument type: tidy-eval}
\description{
This page the argument modifier \verb{<tidy-eval>}. This indicates that the
argument uses \strong{tidy evaluation} with \strong{data masking}; a special type of
non-standard evaluation used throughout the tidyverse.
}
\keyword{internal}
\section{Key terms}{

The primary motivation for tidy evaluation is that it provides data-masking,
which blurs the distinction between two types of variables:
\itemize{
\item \strong{env-variables} are "programming" variables and live in an environment.
They are usually created with \verb{<-}. Env-variables can be any type of R
object.
\item \strong{data-variables} are "statistical" variables and live in a data frame.
They usually come from data files (e.g. \code{csv}), or are created by
manipulating existing variables. Data-variables live inside data frames,
so must be vectors.
}
}
\section{Referring to variables indirectly}{

There are two main cases:
\itemize{
\item If you have the column name as a character vector, use the \code{.data}
pronoun: \code{.data[[var]]}.
\item If you want the user to supply the colum name in a function call,
\strong{embrace} the argument name: \code{{{ var }}}.
}
}
