% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr_dplyr}
\alias{dr_dplyr}
\title{Dr Dplyr checks your installation for common problems.}
\usage{
dr_dplyr()
}
\description{
Only run this if you are seeing problems, like random crashes.
It's possible for \code{dr_dplyr} to return false positives, so there's no
need to run if all is ok.
}
\examples{
\dontrun{
dr_dplyr()
}
}
