% Generated by roxygen2 (4.0.0): do not edit by hand
\name{explain}
\alias{explain}
\title{Explain details of an object}
\usage{
explain(x, ...)
}
\arguments{
\item{x}{An object to explain}

\item{...}{Other parameters possibly used by generic}
}
\description{
This is a generic function which gives more details about an object than
\code{\link{print}}, and is more focussed on human readable output than
\code{\link{str}}.
}
\details{
For more details of explaining dplyr sql tbls, see \code{\link{explain_tbl}}.
}

