% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_n.R
\name{first_n}
\alias{first_n}
\title{Get the first n rows of each group}
\usage{
first_n(.data, n)
}
\arguments{
\item{.data}{a tbl}

\item{n}{if positive, the \code{n} first rows are kept, otherwise the \code{-n} last rows are dropped}
}
\description{
Get the first n rows of each group
}
\examples{

first_n(iris, 3)

# the first 4 rows of each group
iris \%>\%
  group_by(Species) \%>\%
  first_n(4)

# Keep n() - 49 rows for each group
iris \%>\%
  group_by(Species) \%>\%
  first_n(-49)
}
