% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-df.r, R/utils-format.r
\name{print.tbl_df}
\alias{dim_desc}
\alias{formatting}
\alias{print.tbl_df}
\alias{trunc_mat}
\title{Tools for describing matrices}
\usage{
\method{print}{tbl_df}(x, ..., n = NULL, width = NULL)

dim_desc(x)

trunc_mat(x, n = NULL, width = NULL, n_extra = 100)
}
\arguments{
\item{x}{Object to show.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print
all rows if less than option \code{dplyr.print_max}. Otherwise, will
print \code{dplyr.print_min}}

\item{width}{Width of text output to generate. This defaults to NULL, which
means use \code{getOption("width")} and only display the columns that
fit on one screen. You can also set \code{options(dplyr.width = Inf)} to
override this default and always print all columns.}
}
\description{
Tools for describing matrices
}
\examples{
dim_desc(mtcars)
trunc_mat(mtcars)

print(tbl_df(mtcars))
print(tbl_df(mtcars), n = 1)
print(tbl_df(mtcars), n = 3)
print(tbl_df(mtcars), n = 100)

}
\keyword{internal}

