% Generated by roxygen2 (4.0.0): do not edit by hand
\name{funs}
\alias{funs}
\title{Create a list of functions.}
\usage{
funs(..., env = parent.frame())
}
\arguments{
\item{...}{You can specify functions by:

  \itemize{
    \item Their name, \code{"mean"}
    \item The function itself, \code{mean}
    \item A call to the function with \code{.} as a dummy parameter,
      \code{mean(., na.rm = TRUE)}
  }}
}
\description{
\code{funs} provides a flexible to generate a named list of functions.
}
\examples{
funs(mean, "mean", mean(., na.rm = TRUE))

# Overide default names
funs(m1 = mean, m2 = "mean", m3 = mean(., na.rm = TRUE))
}

