\name{group_by}
\alias{group_by}
\title{Group a data source by one or more variables.}
\usage{
  group_by(x, ...)
}
\arguments{
  \item{x}{a data source}

  \item{...}{variables to group by. All data sources accept
  variable names, some will also accept functons of
  variables.}
}
\description{
  Group a data source by one or more variables.
}

