% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise-group-by.R
\name{group_by_all}
\alias{group_by_all}
\alias{group_by_at}
\alias{group_by_if}
\title{Group by a selection of variables.}
\usage{
group_by_all(.tbl)

group_by_at(.tbl, .vars, .add = FALSE)

group_by_if(.tbl, .predicate, .add = FALSE)
}
\arguments{
\item{.tbl}{A \code{tbl} object.}

\item{.vars}{A list of columns generated by \code{\link[=vars]{vars()}},
or a character vector of column names, or a numeric vector of column
positions.}

\item{.add}{Passed to the \code{add} argument of \code{\link[=group_by]{group_by()}}.}

\item{.predicate}{A predicate function to be applied to the columns
or a logical vector. The variables for which \code{.predicate} is or
returns \code{TRUE} are selected. This argument is passed to
\code{\link[rlang:as_function]{rlang::as_function()}} and thus supports quosure-style lambda
functions and strings representing function names.}
}
\description{
Group by a selection of variables.
}
\examples{
# Group a data frame by all variables:
group_by_all(mtcars)

# Group by variables selected with a predicate:
group_by_if(iris, is.factor)
group_by_if(mtcars, rlang::is_integerish)

# Group by variables selected by name:
group_by_at(mtcars, vars(vs, am))
}
