% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.r
\name{group_by_drop_default}
\alias{group_by_drop_default}
\title{Default value for .drop argument of group_by}
\usage{
group_by_drop_default(.tbl)
}
\arguments{
\item{.tbl}{A data frame}
}
\value{
\code{TRUE} unless \code{.tbl} is a grouped data frame that was previously
obtained by \code{group_by(.drop = FALSE)}
}
\description{
Default value for .drop argument of group_by
}
\examples{
group_by_drop_default(iris)

iris \%>\%
  group_by(Species) \%>\%
  group_by_drop_default()

iris \%>\%
  group_by(Species, .drop = FALSE) \%>\%
  group_by_drop_default()

}
\keyword{internal}
