% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct.R, R/group-by.R
\name{distinct_prepare}
\alias{distinct_prepare}
\alias{group_by_prepare}
\title{Prepare for grouping and other operations}
\usage{
distinct_prepare(
  .data,
  vars,
  group_vars = character(),
  .keep_all = FALSE,
  caller_env = caller_env(2),
  error_call = caller_env()
)

group_by_prepare(
  .data,
  ...,
  .add = FALSE,
  .dots = deprecated(),
  add = deprecated(),
  error_call = caller_env()
)
}
\value{
A list
\item{data}{Modified tbl}
\item{groups}{Modified groups}
}
\description{
\verb{*_prepare()} performs standard manipulation that is needed prior
to actual data processing. They are only be needed by packages
that implement dplyr backends.
}
\keyword{internal}
