% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-indices.R
\name{group_indices}
\alias{group_indices}
\alias{group_indices_}
\title{Group id.}
\usage{
group_indices(.data, ...)

group_indices_(.data, ..., .dots)
}
\arguments{
\item{.data}{a tbl}

\item{...}{variables to group by. All tbls accept variable names,
some will also accept functions of variables. Duplicated groups
will be silently dropped.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\description{
Generate a unique id for each group
}
\examples{
group_indices(mtcars, cyl)
}
\seealso{
\code{\link{group_by}}
}

