% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nest.R
\name{group_nest}
\alias{group_nest}
\alias{group_nest_at}
\alias{group_nest_if}
\title{Nest a tibble using a grouping specification}
\usage{
group_nest(.data, ..., .key = "data")

group_nest_at(.data, ..., .key = "data")

group_nest_if(.data, ..., .key = "data")
}
\arguments{
\item{.data}{a tbl}

\item{...}{Variables to nest by, ignored if \code{.data} is already a grouped_df.}

\item{.key}{the name of the list column}
}
\value{
A tbl with one row per unique combination of the grouping variables. The first columns are the grouping variables,
followed by a list column of tibbles with matching rows of the remaining columns.
}
\description{
\badgeexperimental
}
\details{
When \code{.data} is an ungrouped data frame, the \code{...} specifiy the columns to group \emph{by} are are forwarded
to either \code{\link[=group_by]{group_by()}}, \code{\link[=group_by_at]{group_by_at()}} or \code{\link[=group_by_if]{group_by_if()}}.

When \code{.data} is a grouped data frame already, its grouping is used. A warning is issued if \code{...}
is used with a grouped data frame.
}
\examples{

# using group_nest() or its column wise wariants on a ungrouped data frame
starwars \%>\%
  group_nest(species, homeworld)

starwars \%>\%
  group_nest_at(vars(ends_with("_color")))

starwars \%>\%
  group_by_if(is.numeric)

# using group_nest() on a grouped data frame
starwars \%>\%
  group_by(species, homeworld) \%>\%
  group_nest()

# can be useful when the grouped data has been altered before the nesting
# in this example the output is nested by homeworld after the summarise
# has peeled off one level of grouping
starwars \%>\%
  group_by(homeworld, species) \%>\%
  summarise(n = n()) \%>\%
  group_nest()

}
\concept{grouping functions

Conceptually, `group_nest` is the same as `group_by()` + `tidyr::nest()`.}
\keyword{internal}
