% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{group_rows}
\alias{group_rows}
\title{List of rows of each group}
\usage{
group_rows(.data)
}
\arguments{
\item{.data}{a tibble}
}
\value{
a list of integer vectors, giving the indices of the rows of each group
}
\description{
List of rows of each group
}
\examples{
df <- tibble(x = c(1,1,2,2))
group_rows(df)
group_rows(group_by(df, x))
}
