% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike.R
\name{hike}
\alias{hike}
\title{Run expression in each group}
\usage{
hike(.data, ...)
}
\arguments{
\item{.data}{A data frame}

\item{expr}{An expression to execute for each group}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
iris \%>\%
  group_by(Species) \%>\%
  hike(
    mod = lm(Petal.Length ~ Sepal.Length, data = across()),
    print(broom::tidy(mod))
  )

}
