% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hybrid_call}
\alias{hybrid_call}
\title{Inspect how dplyr evaluates an expression}
\usage{
hybrid_call(.data, expr)
}
\arguments{
\item{.data}{a tibble}

\item{expr}{an expression}
}
\description{
Inspect how dplyr evaluates an expression
}
\examples{
# hybrid evaulation
hybrid_call(iris, n())

# standard evaluation
hybrid_call(iris, n() + 1L)
}
