% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dbplyr.R
\name{ident}
\alias{ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)
}
\arguments{
\item{...}{A character vector, or name-value pairs.}
}
\description{
\code{ident()} marks strings as database identifiers (e.g. table or column names)
quoting them using the identifier rules for your database. It is used
primarily in \code{\link[dbplyr:translate_sql]{translate_sql()}} to label variables as identifiers; use
elsewhere should be regarded with suspicion.

\code{ident()} is for internal use only; if you need to supply an table name that
is qualified with schema or catalog use \code{I()}.
}
\examples{
\dontshow{if (requireNamespace("dbplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Identifiers are escaped with "
ident("x")
\dontshow{\}) # examplesIf}
}
