% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_logging}
\alias{init_logging}
\title{Enable internal logging}
\usage{
init_logging(log_level)
}
\arguments{
\item{log_level}{A character value, one of "WARN", "INFO", "DEBUG", "VERB",
or "NONE".}
}
\description{
Log entries, depending on the log level, will be printed to the standard
error stream.
}
\keyword{internal}
