% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils-format.r
\name{knit_print.trunc_mat}
\alias{knit_print.trunc_mat}
\title{knit_print method for trunc mat}
\usage{
knit_print.trunc_mat(x, options)
}
\description{
knit_print method for trunc mat
}
\keyword{internal}

