% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-ops.R
\name{lazy_ops}
\alias{lazy_ops}
\alias{op_base}
\alias{op_double}
\alias{op_grps}
\alias{op_single}
\alias{op_sort}
\alias{op_vars}
\title{Lazy operations}
\usage{
op_base(name, src, x, vars)

op_single(name, x, dots = list(), args = list())

op_double(name, x, y, args = list())

op_vars(op)

op_grps(op)

op_vars(op)

op_sort(op)
}
\description{
This set of S3 classes describe the action of dplyr verbs. These are
currently used for SQL sources to separate the description of operations
in R from their computation in SQL. This API is very new so is likely
to evolve in the future.
}
\details{
\code{op_vars} and \code{op_grps} compute the variables and groups from
a sequence of lazy operations. \code{op_sort} tracks the order of the
data for use in window functions.
}
\keyword{internal}

