% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linewise.R
\name{linewise}
\alias{linewise}
\title{EXperimental replacement for rowwise()}
\usage{
linewise(.data, ...)
}
\arguments{
\item{.data}{A tibble}

\item{...}{Expressions}
}
\description{
EXperimental replacement for rowwise()
}
\examples{
mtcars \%>\%
group_by(cyl) \%>\%
 summarise(fit = list(lm(gear ~ carb))) \%>\%
 linewise(broom::tidy(fit))
}
