% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.r
\name{make_tbl}
\alias{make_tbl}
\title{Create a "tbl" object}
\usage{
make_tbl(subclass, ...)
}
\arguments{
\item{subclass}{name of subclass. "tbl" is an abstract base class, so you
must supply this value. \code{tbl_} is automatically prepended to the
class name}

\item{...}{For \code{tbl()}, other fields used by class. For \code{as.tbl()},
other arguments passed to methods.}

\item{object}{to test/coerce.}
}
\description{
\code{tbl()} is the standard constructor for tbls. \code{as.tbl()} coerces,
and \code{is.tbl()} tests.
}
\keyword{internal}
