% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-sqlite.r
\name{memdb_frame}
\alias{memdb_frame}
\title{Create a database table in temporary in-memory database.}
\usage{
memdb_frame(..., .name = random_table_name())
}
\arguments{
\item{...}{A set of name-value pairs. Arguments are evaluated sequentially,
so you can refer to previously created variables.}

\item{.name}{Name of table in database: defaults to a random name that's
unlikely to conflict with an existing table.}
}
\description{
\code{memdb_frame()} works like \code{\link[tibble:tibble]{tibble::tibble()}}, but instead of creating a new
data frame in R, it creates a table in \code{\link[=src_memdb]{src_memdb()}}.
}
\examples{
if (require("RSQLite")) {
df <- memdb_frame(x = runif(100), y = runif(100))
df \%>\% arrange(x)
df \%>\% arrange(x) \%>\% show_query()
}

}
