% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Move variables}
\usage{
move(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link[=tbl_df]{tbl_df()}}, \code{\link[dtplyr:tbl_dt]{dtplyr::tbl_dt()}} and \code{\link[dbplyr:tbl_dbi]{dbplyr::tbl_dbi()}}.}

\item{...}{A selection of variables to move. Passed to \code{\link[=select]{select()}}.}

\item{.before, .after}{A selection for a variable to use as the reference
location. Exactly one of these arguments must be supplied.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Move a selection of variables in a tbl to a location relative to another
variable.
}
\details{
The variable in the reference location will not be moved, even if it is
selected with \code{...}. If more than one variable is picked in the reference
location selection, \code{.before} uses the first selected variable as the
reference location, and \code{.after} uses the last.
}
\examples{
iris <- as_tibble(iris) # so it prints a little nicer
move(iris, Petal.Length, .after = Sepal.Length)
move(iris, Sepal.Width, .before = Petal.Width)

# Move many variables
move(iris, starts_with("Sepal"), .before = Species)
move(iris, starts_with("Sepal"), .after = starts_with("Petal"))

# Move a variable to first
move(iris, Species, .before = everything())
move(iris, everything(), .after = Species)

# Move a variable to last
move(iris, Sepal.Length, .after = last_col())
move(iris, everything(), .before = Sepal.Length)
}
\seealso{
Other single table verbs: \code{\link{arrange}},
  \code{\link{filter}}, \code{\link{mutate}},
  \code{\link{select}}, \code{\link{slice}},
  \code{\link{summarise}}
}
\concept{single table verbs}
