% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{n}
\alias{n}
\title{The number of observations in the current group.}
\usage{
n()
}
\description{
This function is implemented special for each data source and can only
be used from within \code{\link{summarise}}, \code{\link{mutate}} and
\code{\link{filter}}
}
\examples{
if (require("nycflights13")) {
carriers <- group_by(flights, carrier)
summarise(carriers, n())
mutate(carriers, n = n())
filter(carriers, n() < 100)
}
}

