% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{named_commas}
\alias{named_commas}
\title{Provides comma-separated string out ot the parameters}
\usage{
named_commas(...)
}
\arguments{
\item{...}{Arguments to be constructed into the string}
}
\description{
Provides comma-separated string out ot the parameters
}
\keyword{internal}

