% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_by.R
\name{nest_by}
\alias{nest_by}
\alias{nest_by_at}
\alias{nest_by_if}
\title{Nest by one or more variables}
\usage{
nest_by(.data, ..., .key = "data")

nest_by_at(.data, ..., .key = "data")

nest_by_if(.data, ..., .key = "data")
}
\arguments{
\item{.data}{a tbl}

\item{...}{Variables to nest by. \code{nest_by} passes them to \code{\link[=group_by]{group_by()}}, \code{nest_by_at} passes them to \code{\link[=group_by_at]{group_by_at()}} and \code{nest_by_if()} passes them to \code{\link[=group_by_if]{group_by_if()}}}

\item{.key}{the name of the list column}
}
\value{
A tbl with one row per unique combination of the grouping variables. The first columns are the grouping variables,
followed by a list column of tibbles with matching rows of the remaining columns.
}
\description{
\badgeexperimental
}
\details{
Conceptually, \code{nest_by} is the same as \code{group_by()} + \code{tidyr::nest()}.
}
\examples{
starwars \%>\%
  nest_by(species, homeworld)

starwars \%>\%
  nest_by_at(vars(ends_with("_color")))

starwars \%>\%
  nest_by_if(is.numeric)
}
