% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped-df.r
\name{new_grouped_df}
\alias{new_grouped_df}
\title{Create a custom grouped tibble}
\usage{
new_grouped_df(.data, group_data, class = NULL)
}
\arguments{
\item{.data}{A data frame}

\item{group_data}{The grouped structure, see details}

\item{class}{An additional class for the resulting tibble}
}
\description{
Create a custom grouped tibble
}
\details{
\code{group_data} should be a data frame. Its last column should be called \code{.rows} and be
a list of 1 based integer vectors
}
\examples{
# 5 bootstrap samples
tbl <- new_grouped_df(
  tibble(x = rnorm(10)),
  tibble(.rows = replicate(5, sample(1:10, replace = TRUE), simplify = FALSE))
)
# mean of each bootstrap sample
summarise(tbl, x = mean(x))

}
