% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nycflights13.r
\name{nycflights13}
\alias{copy_nycflights13}
\alias{has_nycflights13}
\alias{nycflights13}
\alias{nycflights13_postgres}
\alias{nycflights13_sqlite}
\title{Database versions of the nycflights13 data}
\usage{
nycflights13_sqlite(path = NULL)

nycflights13_postgres(dbname = "nycflights13", ...)

has_nycflights13(type = c("sqlite", "postgresql"), ...)

copy_nycflights13(src, ...)
}
\arguments{
\item{path}{location of sqlite database file}

\item{dbname, ...}{Arguments passed on to \code{\link{src_postgres}}}
}
\description{
These functions cache the data from the \code{nycflights13} database in
a local database, for use in examples and vignettes. Indexes are created
to making joining tables on natural keys efficient.
}
\keyword{internal}

