% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{peek_mask}
\alias{peek_mask}
\title{Current instance of dplyr's DataMask}
\usage{
peek_mask(fun = "peek_mask()")
}
\arguments{
\item{fun}{Name of the caller function, only used by error messages}
}
\value{
A DataMask object
}
\description{
Current instance of dplyr's DataMask
}
