% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.r
\name{put}
\alias{put}
\title{Assign to a name after a chain of data manipulations}
\usage{
put(x, name, where = NULL)
}
\arguments{
\item{x}{a value to be assigned to \code{name}.}

\item{name}{an unquoted variable name.}

\item{where}{where to do the assignment. Default is the environment in which
the chain of manipulations is called. Passed to \code{pos} argument of
\code{\link{assign}}.}
}
\value{
The function is invoked for its side effect, which is to assign
  \code{x} to the variable \code{name}. \code{x} is then returned.
}
\description{
Alias for the base function \code{\link{assign}}, which makes it easier to
assign the output of a chain of data manipulations.
}
\examples{
mtcars \%>\%
group_by(cyl) \%>\%
summarise_each(funs(mean), drat, wt, qsec) \%>\%
put(car_cyls)

print(car_cyls)

# It is also possible to assign in the middle of a chain.
mtcars \%>\%
group_by(cyl) \%>\%
put(cars_by_cyl) \%>\%
summarise_each(funs(mean, sd), drat, wt, qsec) \%>\%
right_join(cars_by_cyl) \%>\%
put(mtcars_aug)
ls()

print(cars_by_cyl)
print(mtcars_aug)

}

