% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query}
\alias{query}
\title{Create a mutable query object.}
\usage{
query(con, sql, .vars)
}
\arguments{
\item{con}{a \link[DBI:DBIConnection-class]{DBI::DBIConnection} object}

\item{sql}{a string containing an sql query.}
}
\description{
A query object is mutable wrapper around a \linkS4class{DBIResult} object that caches
expensive operations, and insulates the rest of dplyr from the vagaries of
DBI and the individual database implementation.
}
\keyword{internal}
