% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-vars.R
\name{quo_is_select_helper}
\alias{quo_is_select_helper}
\title{Determine if an expression quotation is a helper function}
\usage{
quo_is_select_helper(quo)
}
\arguments{
\item{quo}{A \code{quo} object.}
}
\value{
A logical value
}
\description{
\code{quo_is_select_helper} takes a \code{quo} object and returns a logical indicating if the expression contains certain calls.
}
\examples{
quo_is_select_helper(quo(a_variable_name))
quo_is_select_helper(quo(everything()))

}
