% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataframe.R
\name{recycle}
\alias{recycle}
\alias{recycle_each}
\alias{recycle_whole}
\title{Recycle a Vector}
\usage{
recycle(x, n = NULL)

recycle_each(x, times = (n/length(x)), n = NULL)

recycle_whole(x, n = NULL)
}
\arguments{
\item{x}{An (atomic, one-dimensional) vector.}

\item{n}{The desired output length. Note that, when
  evaluated within the context of a \code{data_frame}
  call, \code{n} will automatically be filled with
  the resultant number of rows.}

\item{times}{The number of times the vector (or, in the case of
\code{recycle_each}, each element of the vector) should be
repeated.}
}
\description{
Helper functions (primarily for \code{\link{data_frame}})
that enable explicit recycling for vectors.
}
\examples{
recycle_each(1:5, 1:5, 30)
recycle(1:2, 10)

# Best used in conjunction with \\code{data_frame}, where the
# output length is automatically inferred for you:

data_frame(x = 1:20, y = recycle_each(1:4, 1:4))
}

