% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-magrittr.R, R/reexport-pillar.R,
%   R/reexport-tibble.R, R/select-helpers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{type_sum}
\alias{data_frame}
\alias{as_data_frame}
\alias{lst}
\alias{add_row}
\alias{tribble}
\alias{tibble}
\alias{as_tibble}
\alias{view}
\alias{contains}
\alias{select_helpers}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{last_col}
\alias{any_of}
\alias{all_of}
\alias{where}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{pillar}{\code{\link[pillar:type_sum]{type_sum()}}}

  \item{tibble}{\code{\link[tibble:add_row]{add_row()}}, \code{\link[tibble:as_data_frame]{as_data_frame()}}, \code{\link[tibble:as_tibble]{as_tibble()}}, \code{\link[tibble:data_frame]{data_frame()}}, \code{\link[tibble:lst]{lst()}}, \code{\link[tibble:tibble]{tibble()}}, \code{\link[tibble:tribble]{tribble()}}, \code{\link[tibble:view]{view()}}}

  \item{tidyselect}{\code{\link[tidyselect:all_of]{all_of()}}, \code{\link[tidyselect:any_of]{any_of()}}, \code{\link[tidyselect:contains]{contains()}}, \code{\link[tidyselect:ends_with]{ends_with()}}, \code{\link[tidyselect:everything]{everything()}}, \code{\link[tidyselect:last_col]{last_col()}}, \code{\link[tidyselect:matches]{matches()}}, \code{\link[tidyselect:num_range]{num_range()}}, \code{\link[tidyselect:one_of]{one_of()}}, \code{\link[tidyselect:starts_with]{starts_with()}}, \code{\link[tidyselect:where]{where()}}}
}}

