% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{rows_truncate}
\alias{rows_truncate}
\title{rows_truncate}
\usage{
rows_truncate(x, ..., copy = FALSE, in_place = FALSE)
}
\arguments{
\item{x}{A pair of data frames or data frame extensions (e.g. a tibble).}

\item{...}{Other parameters passed onto methods.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{in_place}{Should \code{x} be modified in place? This argument is only
relevant for mutable backends (e.g. databases, data.tables).

When \code{TRUE}, a modified version of \code{x} is returned invisibly;
when \code{FALSE}, a new object representing the resulting changes is returned.}
}
\description{
\code{rows_truncate()} removes all rows.
This operation corresponds to \code{TRUNCATE} in SQL.
\code{...} is ignored.
}
