% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rowwise}
\alias{rowwise}
\title{Group input by rows}
\usage{
rowwise(data)
}
\arguments{
\item{data}{Input data frame.}
}
\description{
\code{rowwise} is used for the results of \code{\link{do}} when you
create list-variables. It is also useful to support arbitrary
complex operations that need to be applied to each row.
}
\details{
Currently \code{rowwise} grouping only works with data frames.
}
\examples{
df <- expand.grid(x = 1:3, y = 3:1)
# df \%>\% rowwise() \%>\% do(seq(.$x, .$y))
}

