% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-tidyselect.R
\name{select_vars}
\alias{select_vars}
\alias{rename_vars}
\alias{select_var}
\alias{current_vars}
\title{Select variables}
\usage{
select_vars(vars = chr(), ..., include = chr(), exclude = chr())

rename_vars(vars = chr(), ..., strict = TRUE)

select_var(vars, var = -1)

current_vars(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions now live in the tidyselect package as
\code{\link[tidyselect:vars_select]{tidyselect::vars_select()}}, \code{\link[tidyselect:vars_select]{tidyselect::vars_rename()}} and
\code{\link[tidyselect:vars_pull]{tidyselect::vars_pull()}}.
}
\keyword{internal}
