% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-utils.R
\name{set_current_vars}
\alias{set_current_vars}
\title{Select variables in an environment}
\usage{
set_current_vars(x)
}
\arguments{
\item{x}{A character vector of variable names.}
}
\value{
Invisibly returns the original version of the environment
prior to selecting the variables in \code{x}.
}
\description{
\code{set_current_vars} marks a set of variables as selected in an
environment.
}
