% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.r
\name{setops}
\alias{setops}
\alias{intersect}
\alias{union}
\alias{union_all}
\alias{setdiff}
\alias{setequal}
\title{Set operations}
\usage{
intersect(x, y, ...)

union(x, y, ...)

union_all(x, y, ...)

setdiff(x, y, ...)

setequal(x, y, ...)
}
\arguments{
\item{x, y}{Pair of data frames.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Perform set operations using the rows of a data frame or tibble.
\code{intersect()}, \code{union()}, \code{setdiff()}, and \code{setequal()} override the base
functions of the same name in order to make them generic. The existing
behaviour for vectors is preserved by providing default methods that call
the base functions.

Note that \code{intersect()}, \code{union()} and \code{setdiff()} remove duplicates
in \code{x} and \code{y}.
}
\examples{
mtcars$model <- rownames(mtcars)
first <- mtcars[1:20, ]
second <- mtcars[10:32, ]

intersect(first, second)
union(first, second)
setdiff(first, second)
setdiff(second, first)

union_all(first, second)
setequal(mtcars, mtcars[32:1, ])

# Note these functions also remove pre-existing duplicates in x or y:
a <- data.frame(x = c(1:3, 3, 3))
b <- data.frame(x = c(3:5, 5))

intersect(a, b)
union(a, b)
setdiff(a, b)
}
