% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{slice}
\alias{slice}
\alias{slice_}
\title{Select rows by position.}
\usage{
slice(.data, ...)

slice_(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link{tbl_df}}, \code{\link{tbl_dt}} and \code{\link{tbl_sql}}.}

\item{...}{Integer row values}
}
\description{
Select rows by position.
}
\examples{
slice(mtcars, 1L)
slice(mtcars, n())
slice(mtcars, 5:n())

by_cyl <- group_by(mtcars, cyl)
slice(by_cyl, 1:2)
}
\seealso{
Other single.table.verbs: \code{\link{arrange}},
  \code{\link{arrange_}}; \code{\link{filter}};
  \code{\link{mutate}}, \code{\link{mutate_}},
  \code{\link{transmute}}; \code{\link{rename}},
  \code{\link{select}}, \code{\link{select_}};
  \code{\link{summarise}}, \code{\link{summarise_}},
  \code{\link{summarize}}, \code{\link{summarize_}}
}

