% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by.R
\name{split_by}
\alias{split_by}
\alias{split_by_at}
\alias{split_by_if}
\title{Split data frame by groups}
\usage{
split_by(.data, ...)

split_by_at(.data, ...)

split_by_if(.data, ...)
}
\arguments{
\item{.data}{A data frame}

\item{...}{See \code{\link[=group_by]{group_by()}}}
}
\value{
a list of tibbles
}
\description{
Split data frame by groups
}
\examples{

}
