% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dbplyr.R
\name{sql}
\alias{sql}
\title{Literal SQL escaping}
\usage{
sql(...)
}
\arguments{
\item{...}{Character vectors that will be combined into a single SQL vector.}
}
\description{
Use \code{sql()} to declare that a string is literal SQL and should be used
as is, without quoting.
}
