% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-lazy.R
\name{sql_build}
\alias{select_query}
\alias{sql_build}
\alias{sql_render}
\title{Build and render SQL from a sequence of lazy operations}
\usage{
sql_build(op, ...)

select_query(from, select, where = character(), group_by = character(),
  having = character(), order_by = character(), limit = NULL,
  offset = NULL)

sql_render(x, con = NULL, ...)
}
\description{
Build creates a \code{select_query} S3 object, and then \code{sql_render}
converts it into a SQL string.
}
\keyword{internal}

