% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-escape.r
\name{sql_quote}
\alias{sql_quote}
\title{Helper function for quoting sql elements.}
\usage{
sql_quote(x, quote)
}
\arguments{
\item{x}{Character vector to escape.}

\item{quote}{Single quoting character.}
}
\description{
If the quote character is present in the string, it will be doubled.
\code{NA}s will be replaced with NULL.
}
\examples{
sql_quote("abc", "'")
sql_quote("I've had a good day", "'")
sql_quote(c("abc", NA), "'")
}
\keyword{internal}

