% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-test.r
\name{src-test}
\alias{db_query_fields.DBITestConnection}
\alias{sql_escape_ident.DBITestConnection}
\alias{sql_translate_env.DBITestConnection}
\alias{src-test}
\title{A set of DBI methods to ease unit testing dplyr with DBI}
\usage{
\method{db_query_fields}{DBITestConnection}(con, sql, ...)

\method{sql_escape_ident}{DBITestConnection}(con, x)

\method{sql_translate_env}{DBITestConnection}(con)
}
\arguments{
\item{con}{A database connection.}

\item{sql}{a string containing an sql query.}

\item{...}{Other arguments passed on to the individual methods}

\item{x}{Object to transform}
}
\description{
A set of DBI methods to ease unit testing dplyr with DBI
}

