% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-storms.R
\docType{data}
\name{storms}
\alias{storms}
\title{Storm tracks data}
\format{
A tibble with 22,184 observations and 13 variables:
\describe{
\item{name}{Storm Name}
\item{year,month,day}{Date of report}
\item{hour}{Hour of report (in UTC)}
\item{lat,long}{Location of storm center}
\item{status}{Storm classification (Tropical Depression, Tropical Storm,
or Hurricane)}
\item{category}{Saffir-Simpson hurricane category calculated from wind speed. Only applies to hurricanes.
\itemize{
\item 64+ knots = Category 1
\item 83+ knots = Category 2
\item 96+ knots = Category 3
\item 113+ knots = Category 4
\item 137+ knots = Category 5
}
}
\item{wind}{storm's maximum sustained wind speed (in knots)}
\item{pressure}{Air pressure at the storm's center (in millibars)}
\item{tropicalstorm_force_diameter}{Diameter (in nautical miles) of the area experiencing tropical storm strength winds (34 knots or above). Only available starting in 2004.}
\item{hurricane_force_diameter}{Diameter (in nautical miles) of the area experiencing hurricane strength winds (64 knots or above). Only available starting in 2004.}
}
}
\usage{
storms
}
\description{
This dataset is the NOAA Atlantic hurricane database best track data, \url{https://www.nhc.noaa.gov/data/#hurdat}.
The data includes the positions and attributes of storms from 1852-2021.
The modern storms are measured every six hours during the lifetime of the storm.
}
\examples{

# show a plot of the storm paths in 1975 or later
if (requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)
  storms \%>\%
    filter(year >= 1975) \%>\%
  ggplot() +
    aes(x=long, y=lat, color=paste(year, name)) +
    geom_path() +
    guides(color='none') +
    facet_wrap(~year)
}

storms
}
\seealso{
The script to create the storms data set: \url{https://github.com/tidyverse/dplyr/blob/main/data-raw/storms.R}
}
\keyword{datasets}
