% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{summarise_each}
\alias{summarise_each}
\alias{summarize_each}
\alias{mutate_each}
\title{Summarise and mutate multiple columns.}
\usage{
summarise_each(tbl, funs, ...)

summarize_each(tbl, funs, ...)

mutate_each(tbl, funs, ...)
}
\arguments{
\item{tbl}{a tbl}

\item{funs}{List of function calls, generated by \code{\link[=funs]{funs()}}, or
a character vector of function names.}

\item{...}{Variables to include/exclude in mutate/summarise.
You can use same specifications as in \code{\link[=select]{select()}}. If missing,
defaults to all non-grouping variables.

For standard evaluation versions (ending in \code{_}) these can
be either a list of expressions or a character vector.}
}
\description{
Apply one or more functions to one or more columns. Grouping variables
are always excluded from modification.
}
\details{
\code{mutate_each()} and \code{summarise_each()} are deprecated in favour of
a more featureful family of functions: \code{\link[=mutate_all]{mutate_all()}},
\code{\link[=mutate_at]{mutate_at()}}, \code{\link[=mutate_if]{mutate_if()}}, \code{\link[=summarise_all]{summarise_all()}}, \code{\link[=summarise_at]{summarise_at()}}
and \code{\link[=summarise_if]{summarise_if()}}.
}
