% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-tibble.R
\name{tbl_df}
\alias{tbl_df}
\alias{as.tbl}
\title{Coerce to a tibble}
\usage{
tbl_df(data)

as.tbl(x, ...)
}
\arguments{
\item{data, x}{Object to coerce}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link[tibble:as_tibble]{tibble::as_tibble()}} instead.
}
\keyword{internal}
