% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{tbl_ptype}
\alias{tbl_ptype}
\title{Return a prototype of a tbl}
\usage{
tbl_ptype(.data)
}
\description{
Used in \verb{_if} functions to enable type-based selection even when the data
is lazily generated. Should either return the complete tibble, or if that
can not be computed quickly, a 0-row tibble where the columns are of
the correct type.
}
\keyword{internal}
