% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.r
\name{tbl_vars}
\alias{tbl_vars}
\title{List variables provided by a tbl.}
\usage{
tbl_vars(x, group_vars = TRUE)
}
\arguments{
\item{x}{A \code{tbl} object.}

\item{group_vars}{Whether to return group variables as well.}
}
\description{
List variables provided by a tbl.
}
