% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-temp.r
\name{testing}
\alias{test_frame}
\alias{test_load}
\alias{test_register_src}
\alias{testing}
\title{Infrastructure for testing dplyr}
\usage{
test_register_src(name, src)

test_load(df, name = random_table_name(), srcs = test_srcs$get(),
  ignore = character())

test_frame(..., srcs = test_srcs$get())
}
\description{
Register testing sources, then use \code{test_load} to load an existing
data frame into each source. To create a new table in each source,
use \code{test_frame}.
}
\examples{
\dontrun{
test_register_src("df", src_df(env = new.env()))
test_register_src("sqlite", src_sqlite("file::memory:", create = TRUE))

test_frame(x = 1:3, y = 3:1)
test_load(mtcars)
}
}
\keyword{internal}

