% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{vars}
\alias{vars}
\title{Select variables}
\usage{
vars(...)
}
\arguments{
\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Variables to operate on.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{vars()} is superseded because it is only needed for the scoped verbs (i.e.
\code{\link[=mutate_at]{mutate_at()}}, \code{\link[=summarise_at]{summarise_at()}}, and friends), which have been been
superseded in favour of \code{\link[=across]{across()}}. See \code{vignette("colwise")} for details.

This helper is intended to provide tidy-select semantics for scoped verbs
like \code{mutate_at()} and \code{summarise_at()}. Note that anywhere you can supply
\code{vars()} specification, you can also supply a numeric vector of column
positions or a character vector of column names.
}
\seealso{
\code{\link[=all_vars]{all_vars()}} and \code{\link[=any_vars]{any_vars()}} for other quoting
functions that you can use with scoped verbs.
}
