% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_order.R
\name{window_order}
\alias{window_order}
\title{Override window order and frame}
\usage{
window_order(.data, ...)
}
\arguments{
\item{...}{Variables to order by}
}
\description{
These allow you to override the \verb{PARTITION BY} and \verb{ORDER BY} clauses
of window functions generated by grouped mutates.
}
\examples{
library(dbplyr)
library(dplyr, warn.conflicts = FALSE)

set.seed(2022)
db <- memdb_frame(g = rep(1:2, each = 5), y = runif(10), z = 1:10)
db \%>\%
  window_order(y) \%>\%
  mutate(z = cumsum(y))

set.seed(2022)
df <- tibble(g = rep(1:2, each = 5), y = runif(10), z = 1:10)
df \%>\%
  window_order(y) \%>\%
  mutate(z = cumsum(y))

}
