% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_rownames}
\alias{add_rownames}
\title{Convert row names to an explicit variable.}
\usage{
add_rownames(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame with rownames.}

\item{var}{Name of variable to use}
}
\description{
Convert row names to an explicit variable.
}
\examples{
mtcars \%>\%
  tbl_df() \%>\%
  print() \%>\%
  add_rownames()
}

