% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-tibble.R
\name{add_rownames}
\alias{add_rownames}
\title{Convert row names to an explicit variable.}
\usage{
add_rownames(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame with rownames.}

\item{var}{Name of variable to use}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link[tibble:rownames_to_column]{tibble::rownames_to_column()}} instead.
}
\keyword{internal}
