\name{as.data.table.source_dt}
\alias{as.data.table.source_dt}
\title{Coerce data table to source.}
\usage{
  as.data.table.source_dt(x, keep.rownames = NULL)

  as.data.table.source_dt(x, keep.rownames = NULL)
}
\description{
  Coerce data table to source.

  Coerce data table to source.
}
\keyword{internal}

