% Generated by roxygen2 (4.0.2): do not edit by hand
\name{backend_sql}
\alias{backend_sql}
\alias{sql_escape_ident}
\alias{sql_escape_string}
\alias{sql_join}
\alias{sql_select}
\alias{sql_semi_join}
\alias{sql_set_op}
\alias{sql_subquery}
\title{SQL generation.}
\usage{
sql_select(con, select, from, where = NULL, group_by = NULL,
  having = NULL, order_by = NULL, limit = NULL, offset = NULL, ...)

sql_subquery(con, sql, name = random_table_name(), ...)

sql_join(con, x, y, type = "inner", by = NULL, ...)

sql_semi_join(con, x, y, anti = FALSE, by = NULL, ...)

sql_set_op(con, x, y, method)

sql_escape_string(con, x)

sql_escape_ident(con, x)
}
\arguments{
\item{con}{A database connection.}
}
\value{
An SQL string.
}
\description{
These generics are used to run build various SQL queries.  Default methods
are provided for \code{DBIConnectdion}, but variations in SQL across
databases means that it's likely that a backend will require at least a
few methods.
}
\keyword{internal}

