% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbi-s3.r
\name{backend_src}
\alias{backend_src}
\alias{src_desc}
\alias{src_translate_env}
\title{Source generics.}
\usage{
src_desc(x)

src_translate_env(x)
}
\description{
These generics retrieve metadata for a given src.
}
\keyword{internal}

