% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{between}
\alias{between}
\title{Do values in a numeric vector fall in specified range?}
\usage{
between(x, left, right)
}
\arguments{
\item{x}{A numeric vector of values}

\item{left, right}{Boundary values}
}
\description{
This is a shortcut for \code{x >= left & x <= right}, implemented
efficiently in C++ for local values, and translated to the
appropriate SQL for remote tables.
}
\examples{
between(1:12, 7, 9)

x <- rnorm(1e2)
x[between(x, -1, 1)]
}
