% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-rows.R
\name{bind_rows}
\alias{bind_rows}
\alias{bind}
\title{Bind multiple data frames by row}
\usage{
bind_rows(..., .id = NULL)
}
\arguments{
\item{...}{Data frames to combine. Each argument can either be a data frame,
a list that could be a data frame, or a list of data frames. Columns are
matched by name, and any missing columns will be filled with \code{NA}.}

\item{.id}{The name of an optional identifier column. Provide a string to
create an output column that identifies each input. The column will use
names if available, otherwise it will use positions.}
}
\value{
A data frame the same type as the first element of \code{...}.
}
\description{
Bind any number of data frames by row, making a longer result. This is
similar to \code{do.call(rbind, dfs)}, but the output will contain all columns
that appear in any of the inputs.
}
\section{Methods}{

This function is a \strong{generic}, which means that packages can provide
implementations (methods) for other classes. See the documentation of
individual methods for extra arguments and differences in behaviour.

Methods available in currently loaded packages:
\Sexpr[stage=render,results=rd]{dplyr:::methods_rd("bind_rows")}.

\code{bind_rows()} is a special S3 generic due to the fact that its first
argument is \code{...}. S3 methods for \code{bind_rows()} are written as:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{#' @importFrom dplyr bind_rows
#' @export
bind_rows.my_subclass <- function(..., .id = NULL) \{
\}
}\if{html}{\out{</div>}}

Dispatch is performed on the \emph{first} argument, so order matters when data
frames of different types are being combined.

For S3 methods authors:
\itemize{
\item \code{...} will always contain at least 1 input. The first input will have the
S3 class that dispatch was performed on, but no other inputs are validated.
\item \code{...} must be captured by \code{\link[rlang:list2]{rlang::list2()}}.
\item \code{bind_rows()}'s S3 generic will preprocess \code{...} by applying all legacy
flattening behavior and removing any \code{NULL} values. Dispatch occurs on the
first argument remaining after this preprocessing, and S3 methods should
expect that \code{...} contains a flat list of objects to bind.
\item \code{bind_rows()} S3 method signatures must exactly match the signature of the
\code{bind_rows()} generic. Due to how this S3 generic works, additional method
specific arguments are not allowed.
\item If you call \code{rlang::abort()} from your S3 method, you will likely need to
provide \code{abort(call = call("bind_rows"))}, otherwise an intermediate
function may be reported as the error call site.
}
}

\examples{
df1 <- tibble(x = 1:2, y = letters[1:2])
df2 <- tibble(x = 4:5, z = 1:2)

# You can supply individual data frames as arguments:
bind_rows(df1, df2)

# Or a list of data frames:
bind_rows(list(df1, df2))

# When you supply a column name with the `.id` argument, a new
# column is created to link each row to its original data frame
bind_rows(list(df1, df2), .id = "id")
bind_rows(list(a = df1, b = df2), .id = "id")
}
