% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dbplyr.R
\name{check_dbplyr}
\alias{check_dbplyr}
\alias{wrap_dbplyr_obj}
\title{dbplyr compatiblity functions}
\usage{
check_dbplyr()

wrap_dbplyr_obj(obj_name)
}
\description{
In dplyr 0.6.0, a number of databse and SQL functions moved from dplyr to
dbplyr. The generic functions stayed in dplyr (since there is no easy way
to conditionally import a generic from different packages), but many other
SQL and database helper functions moved. If you have written a backend,
these functions generate the code you need to work with both dplyr 0.5.0
dplyr 0.6.0.
}
\examples{
if (requireNamespace("dbplyr", quietly = TRUE)) {
wrap_dbplyr_obj("build_sql")
wrap_dbplyr_obj("base_agg")
}
}
\keyword{internal}
