% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combine_vars}
\alias{combine_vars}
\title{combine_vars}
\usage{
combine_vars(vars, xs)
}
\arguments{
\item{vars}{a character vector of variable names.}

\item{xs}{a list of integers.}
}
\value{
A named vector with the unique indices that correspond to the names.
}
\description{
Combines variable indicies with a corresponding variable list.
}
\examples{
combine_vars(names(iris), list(1:2, 2:3))
combine_vars(names(iris), list(1))
}
