% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\alias{n}
\alias{current_key}
\alias{cur_group_id}
\alias{current_column}
\title{Context dependent expressions}
\usage{
n()

current_key()

cur_group_id()

current_column()
}
\description{
These functions return information about the "current" group or "current"
variable, so only work inside specific contexts
\itemize{
\item \code{n()} gives the current group size.
\item \code{current_key()} gives the group keys, a tibble with one row and one column
for each grouping variable.
\item \code{cur_group_id()} gives the unique identifier for the current group
\item \code{current_column()} gives the current column being used in \code{\link[=across]{across()}}.
}
}
\examples{
df <- data.frame(g = rep(1:3, 1:3), x = runif(6), y = runif(6))

df \%>\%
  group_by(g) \%>\%
  summarise(n = n())

df \%>\%
  group_by(g) \%>\%
  mutate(across(everything(), ~ paste(current_column(), round(.x, 2))))
}
