% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumall}
\alias{cumall}
\alias{cumany}
\alias{cummean}
\title{Cumulativate versions of any, all, and mean}
\usage{
cumall(x)

cumany(x)

cummean(x)
}
\arguments{
\item{x}{For \code{cumall()} and \code{cumany()}, a logical vector; for
\code{cummean()} an integer or numeric vector}
}
\description{
dplyr adds \code{cumall()}, \code{cumany()}, and \code{cummean()} to complete
R's set of cumulate functions to match the aggregation functions available
in most databases
}
