% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data_frame}
\alias{data_frame}
\title{Make a Data Frame}
\usage{
data_frame(...)
}
\arguments{
\item{...}{A set of named arguments.}
}
\description{
A trimmed down version of \code{\link{data.frame}} that:
}
\details{
1. Evaluates its arguments lazily and in order,
2. Foregoes the option of providing \code{row.names},
3. Foregoes column name checking (except that we confirm names are not empty),
4. Does not convert strings to factors.
}
\examples{
a <- 1:5
data_frame(x = a, y = x ^ 2, z = x + y, a = a + 1, b = a)
}

