% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.r
\name{defunct}
\alias{defunct}
\alias{id}
\alias{failwith}
\alias{funs}
\alias{select_vars}
\alias{rename_vars}
\alias{select_var}
\alias{current_vars}
\title{Defunct functions}
\usage{
# Deprecated in 0.5.0 -------------------------------------

id(.variables, drop = FALSE)

# Deprecated in 0.7.0 -------------------------------------

failwith(default = NULL, f, quiet = FALSE)

# Deprecated in 0.8.* -------------------------------------

funs(..., .args = list())

select_vars(vars = chr(), ..., include = chr(), exclude = chr())

rename_vars(vars = chr(), ..., strict = TRUE)

select_var(vars, var = -1)

current_vars(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These function were all deprecated for at least two years and are now
defunct. Executing them to tell you which function to use instead.
}
\keyword{internal}
