% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.r
\name{defunct}
\alias{defunct}
\alias{id}
\alias{failwith}
\title{Defunct functions}
\usage{
id(.variables, drop = FALSE)

failwith(default = NULL, f, quiet = FALSE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
Executing these functions will tell you which function replaces them.
}
\keyword{internal}
