% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{combine}
\alias{src_mysql}
\alias{src_postgres}
\alias{src_sqlite}
\alias{src_local}
\alias{src_df}
\alias{tbl_df}
\alias{as.tbl}
\alias{add_rownames}
\title{Defunct functions}
\usage{
# Deprecated in 1.0.0 -------------------------------------

combine(...)

src_mysql(
  dbname,
  host = NULL,
  port = 0L,
  username = "root",
  password = "",
  ...
)

src_postgres(
  dbname = NULL,
  host = NULL,
  port = NULL,
  user = NULL,
  password = NULL,
  ...
)

src_sqlite(path, create = FALSE)

src_local(tbl, pkg = NULL, env = NULL)

src_df(pkg = NULL, env = NULL)

tbl_df(data)

as.tbl(x, ...)

add_rownames(df, var = "rowname")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions were deprecated for at least two years before being
made defunct. If there's a known replacement, calling the function
will tell you about it.
}
\keyword{internal}
