% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{distinct}
\alias{distinct}
\alias{distinct_}
\title{Select}
\usage{
distinct(.data, ...)

distinct_(.data, vars = character())
}
\arguments{
\item{.data}{a tbl}

\item{...,vars}{Variables to use when determining uniqueness. If there
  are multiple rows for a given combination of inputs, only the first
  row will be preserved.

  \code{...} should be an unquoted, comma-separated list of variable
  names. \code{vars} can be either a character vector of column names,
  or a list of symbols.}
}
\description{
Retain only unique/distinct variables from an input tbl. This is an
efficient version of \code{\link{unique}}. \code{distinct()} is best-suited
for interactive use, \code{distinct_()} for calling from a function.
}
\examples{
df <- data.frame(
  x = sample(10, 100, rep = TRUE),
  y = sample(10, 100, rep = TRUE)
)
nrow(df)
nrow(distinct(df))
}

