% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.tbl_df}
\alias{dim_desc}
\alias{dplyr-formatting}
\alias{print.tbl_df}
\alias{print.tbl_dt}
\alias{print.tbl_sql}
\alias{trunc_mat}
\title{Tools for describing matrices}
\usage{
\method{print}{tbl_df}(x, ..., n = NULL)

\method{print}{tbl_dt}(x, ..., n = NULL)

\method{print}{tbl_sql}(x, ..., n = NULL)

dim_desc(x)

trunc_mat(x, n = NULL)
}
\arguments{
\item{x}{Object to show.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print
all rows if less than option \code{dplyr.print_max}. Otherwise, will
print \code{dplyr.print_min}}
}
\description{
Tools for describing matrices
}
\examples{
dim_desc(mtcars)
trunc_mat(mtcars)

print(tbl_df(mtcars))
print(tbl_df(mtcars), n = 1)
print(tbl_df(mtcars), n = 3)
print(tbl_df(mtcars), n = 100)
}
\keyword{internal}

