% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-failwith.r
\name{failwith}
\alias{failwith}
\title{Fail with specified value.}
\usage{
failwith(default = NULL, f, quiet = FALSE)
}
\arguments{
\item{default}{default value}

\item{f}{function}

\item{quiet}{all error messages be suppressed?}
}
\value{
a function
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("defunct")}
Please use \code{\link[purrr:possibly]{purrr::possibly()}} instead.
}
\keyword{internal}
