% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filter}
\alias{filter}
\title{Return rows with matching conditions.}
\usage{
filter(.data, ...)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link{tbl_df}}, \code{\link{tbl_dt}} and \code{\link{tbl_sql}}.}

\item{...}{Logical predicates. Multiple conditions are combined with \code{&}.}
}
\value{
An object of the same class as \code{.data}.

  Data frame row names are silently dropped. To preserve, convert to an
  explicit variable.
}
\description{
Return rows with matching conditions.
}
\examples{
filter(mtcars, cyl == 8)
filter(mtcars, cyl < 6)
}
\seealso{
Other single.table.verbs: \code{\link{arrange}};
  \code{\link{mutate}}, \code{\link{transmute}};
  \code{\link{rename}}, \code{\link{select}};
  \code{\link{summarise}}, \code{\link{summarize}}
}

