% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_n.R
\name{first_n}
\alias{first_n}
\title{Get the first n rows of each group}
\usage{
first_n(.data, n)
}
\arguments{
\item{.data}{a tbl}

\item{n}{How many rows}
}
\description{
Get the first n rows of each group
}
\examples{

first_n(iris, 3)

iris \%>\%
  group_by(Species) \%>\%
  first_n(4)

}
