% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{funs}
\alias{funs}
\alias{funs_q}
\title{Create a list of functions calls.}
\usage{
funs(..., env = parent.frame())

funs_q(calls, env = parent.frame())
}
\arguments{
\item{env}{The environment in which to evaluate calls.}

\item{calls,...}{A list of functions specified by:

  \itemize{
    \item Their name, \code{"mean"}
    \item The function itself, \code{mean}
    \item A call to the function with \code{.} as a dummy parameter,
      \code{mean(., na.rm = TRUE)}
  }}
}
\description{
\code{funs} provides a flexible to generate a named list of functions for
input to other functions like \code{colwise}.
}
\examples{
funs(mean, "mean", mean(., na.rm = TRUE))

# Overide default names
funs(m1 = mean, m2 = "mean", m3 = mean(., na.rm = TRUE))

# If you have a function names in a vector, use funs_q
fs <- c("min", "max")
funs_q(fs)
}

