% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct.R, R/group-by.r
\name{distinct_prepare}
\alias{distinct_prepare}
\alias{group_by_prepare}
\title{Same basic philosophy as group_by_prepare(): lazy_dots comes in, list of data and
vars (character vector) comes out.}
\usage{
distinct_prepare(.data, vars, group_vars = character(),
  .keep_all = FALSE)

group_by_prepare(.data, ..., .dots = list(), add = FALSE)
}
\value{
A list
\item{data}{Modified tbl}
\item{groups}{Modified groups}
}
\description{
\code{group_by_prepare()} performs standard operations that should
happen before individual methods process the data. This includes
mutating the tbl to add new grouping columns
and updating the groups (based on add)
}
\details{
\code{distinct_prepare()} is similar, but in preparation of a \code{\link[=distinct]{distinct()}}.
}
\keyword{internal}
