% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.r
\name{group_by_prepare}
\alias{group_by_prepare}
\title{Prepare for grouping.}
\usage{
group_by_prepare(.data, ..., .dots = list(), add = FALSE)
}
\value{
A list
\item{data}{Modified tbl}
\item{groups}{Modified groups}
}
\description{
Performs standard operations that should happen before individual methods
process the data. This includes mutating the tbl to add new grouping columns
and updating the groups (based on add)
}
\keyword{internal}
