% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{group_data}
\alias{group_data}
\title{Grouping data}
\usage{
group_data(.data)
}
\arguments{
\item{.data}{a tibble}
}
\value{
a tibble with one row per group. The last column, always called \code{.rows} is a list of integer vectors
indicating the rows for each group. If \code{.data} is a grouped data frame the first columns
are the grouping variables.
}
\description{
Grouping data
}
\examples{
df <- tibble(x = c(1,1,2,2))

# one row
group_data(df)

# 2 rows, one for each group
group_by(df,x) \%>\% group_data()

# n rows
rowwise(df) \%>\% group_data()
}
