% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_data}
\alias{group_data}
\alias{group_keys}
\alias{group_rows}
\alias{group_indices}
\alias{group_vars}
\alias{groups}
\alias{group_size}
\alias{n_groups}
\title{Grouping metadata}
\usage{
group_data(.data)

group_keys(.tbl, ...)

group_rows(.data)

group_indices(.data, ...)

group_vars(x)

groups(x)

group_size(x)

n_groups(x)
}
\arguments{
\item{.data, .tbl, x}{A data frame or extension (like a tibble or grouped
tibble).}

\item{...}{Use of \code{...} is now deprecated; please use \code{group_by()} first
instead.}
}
\description{
\itemize{
\item \code{group_data()}: data frame that defines the grouping structure.
The last column, always called \code{.rows}, is a list of integer vectors that
gives the location of the rows in each group.
\item \code{group_data()}: just the grouping structure from \code{group_data()}.
\item \code{group_rows()}: just row locations from \code{group_data()}
\item \code{group_indices()}: an integer vector the same length as \code{.data} that
gives the group each row belongs to.
\item \code{group_vars()}: names of grouping variables as character vector
\item \code{groups()}: names of grouping as list of symbols
\item \code{group_size()}: vector giving size of each group
\item \code{n_groups()}: total number of groups
}
}
\examples{
df <- tibble(x = c(1,1,2,2))
group_vars(df)
group_rows(df)
group_data(df)

gf <- group_by(df, x)
group_vars(gf)
group_rows(gf)
group_data(gf)
}
\keyword{internal}
