% Generated by roxygen2 (4.0.2): do not edit by hand
\name{group_indices}
\alias{group_indices}
\alias{group_indices_}
\title{Group id.}
\usage{
group_indices(.data, ...)

group_indices_(.data, ..., .dots, add = FALSE)
}
\arguments{
\item{.data}{a tbl}

\item{...}{variables to group by.}
}
\description{
Generate a unique id for each group
}
\examples{
group_indices(mtcars, cyl)
}
\seealso{
\code{\link{group_by}}
}

