% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_keys.R
\name{group_keys}
\alias{group_keys}
\alias{group_keys_at}
\alias{group_keys_if}
\title{The grouping keys}
\usage{
group_keys(.data, ...)

group_keys_at(.data, ...)

group_keys_if(.data, ...)
}
\arguments{
\item{.data}{A tbl}

\item{...}{See \code{\link[=group_by]{group_by()}}, \code{\link[=group_by_at]{group_by_at()}} or \code{\link[=group_by_if]{group_by_if()}}}
}
\value{
A tibble with one row per group and one column per grouping variable
}
\description{
The grouping keys are a tibble with only the grouping information.
}
\examples{
iris \%>\%
  group_keys(Species)

iris \%>\%
  group_keys_at(vars("Species"))

iris \%>\%
  group_keys_if(is.factor)

iris \%>\%
  group_by(Species) \%>\%
  group_keys()

}
