% Generated by roxygen2 (4.0.1): do not edit by hand
\name{group_size}
\alias{group_size}
\alias{n_groups}
\title{Calculate group sizes.}
\usage{
group_size(x)

n_groups(x)
}
\arguments{
\item{x}{a grouped tbl}
}
\description{
Calculate group sizes.
}
\examples{
data("hflights", package = "hflights")

by_day <- hflights \%>\% group_by(Year, Month, DayofMonth)
n_groups(by_day)
group_size(by_day)

by_dest <- hflights \%>\% group_by(Dest)
n_groups(by_dest)
group_size(by_dest)
}

