% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_split.R
\name{group_split}
\alias{group_split}
\alias{group_split_at}
\alias{group_split_if}
\title{Split data frame by groups}
\usage{
group_split(.data, ...)

group_split_at(.data, ...)

group_split_if(.data, ...)
}
\arguments{
\item{.data}{A tbl}

\item{...}{See \code{\link[=group_by]{group_by()}}, \code{\link[=group_by_at]{group_by_at()}} or \code{\link[=group_by_if]{group_by_if()}}}
}
\value{
a list of tibbles. Each tibble contains the rows of \code{.data} for the associated group.
}
\description{
Split data frame by groups
}
\section{Scoped grouping}{


The scoped variants \code{\link[=group_split_at]{group_split_at()}} and \code{\link[=group_split_if]{group_split_if()}} make it easy to split by a
selection of variables.
}

\examples{
iris \%>\%
  group_split(Species)

iris \%>\%
  group_split_at(vars("Species"))

iris \%>\%
  group_split_if(is.factor)

# group_split() an already grouped data frame
iris \%>\%
  group_by(Species) \%>\%
  group_split()

# this can be useful if the grouped data has been altered
# before the split
iris \%>\%
  group_by(Species) \%>\%
  filter(Sepal.Length > 6) \%>\%
  group_split()

}
\seealso{
Other grouping functions: \code{\link{group_by_all}},
  \code{\link{group_by}}, \code{\link{group_indices}},
  \code{\link{group_keys}}, \code{\link{group_rows}},
  \code{\link{group_size}}, \code{\link{groups}}
}
\concept{grouping functions}
