% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.r
\name{groups}
\alias{groups}
\alias{group_vars}
\title{Return grouping variables}
\usage{
groups(x)

group_vars(x)
}
\arguments{
\item{x}{A \code{\link[=tbl]{tbl()}}}
}
\description{
\code{group_vars()} returns a character vector; \code{groups()} returns a list of
symbols.
}
\examples{
df <- tibble(x = 1, y = 2) \%>\% group_by(x, y)
group_vars(df)
groups(df)
}
