% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike.R
\name{hike}
\alias{hike}
\alias{march}
\title{Run expression in each group}
\usage{
hike(.data, expr)

march(.data, expr)
}
\arguments{
\item{.data}{A data frame}

\item{expr}{An expression to execute for each group}
}
\value{
\itemize{
\item \link{hike} returns the results in a list
\item \link{march} discards the results and returns \code{.data} invisibly
}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
iris \%>\%
  group_by(Species) \%>\%
  march(
    print(head(across(), 2))
  )

iris \%>\%
  group_by(Species) \%>\%
  hike(
    broom::tidy(lm(Petal.Length ~ Sepal.Length, data = across()))
  )

}
