% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dbplyr.R
\name{ident}
\alias{ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)
}
\arguments{
\item{...}{A character vector, or name-value pairs}
}
\description{
This ensures that during SQL generation, the values will be quoted as
identifiers, not as strings. \code{as.ident()} coerces to an identifier,
assuming that escaped SQL already represents an identifier.
}
