% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dbplyr.R
\name{ident}
\alias{ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)
}
\arguments{
\item{...}{A character vector, or name-value pairs}
}
\description{
\code{ident()} takes unquoted strings and flags them as identifiers.
\code{ident_q()} assumes its input has already been quoted, and ensures
it does not get quoted again. This is currently used only for
for \code{schema.table}.
}
\examples{
# Identifiers are escaped with "
\dontshow{if (requireNamespace("dbplyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ident("x")
\dontshow{\}) # examplesIf}
}
