% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dbplyr.R
\name{ident}
\alias{ident}
\title{Flag a character vector as SQL identifiers}
\usage{
ident(...)
}
\arguments{
\item{...}{A character vector, or name-value pairs}
}
\description{
\code{ident()} takes unquoted strings and quotes them for you; \code{ident_q()}
assumes its input has already been quoted.
}
\details{
These two \code{ident} clsases are used during SQL generation to make sure
the values will be quoted as, not as strings.
}
