% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lahman}
\alias{has_lahman}
\alias{lahman}
\alias{lahman_df}
\alias{lahman_dt}
\alias{lahman_monetdb}
\alias{lahman_mysql}
\alias{lahman_postgres}
\alias{lahman_sqlite}
\alias{lahman_srcs}
\title{Cache and retrieve an \code{src_sqlite} of the Lahman baseball database.}
\usage{
lahman_sqlite()

lahman_postgres(...)

lahman_mysql(...)

lahman_monetdb(...)

lahman_df()

lahman_dt()

has_lahman(type, ...)

lahman_srcs(..., quiet = NULL)
}
\arguments{
\item{...}{Arguments passed to \code{src} on first
load. For mysql and postgresql, the defaults assume you have a local
server with \code{lahman} database already created.
For \code{lahman_srcs}, character vector of names giving srcs to generate.}

\item{type}{src type.}

\item{quiet}{if \code{TRUE}, suppress messages about databases failing to
connect.}
}
\description{
This creates an interesting database using data from the Lahman baseball
data source, provided by Sean Lahman at
\url{http://www.seanlahman.com/baseball-archive/statistics/}, and
made easily available in R through the \pkg{Lahman} package by
Michael Friendly, Dennis Murphy and Martin Monkman. See the documentation
for that package for documentation of the inidividual tables.
}
\examples{
# Connect to a local sqlite database, if already created
if (require("RSQLite") && has_lahman("sqlite")) {
  lahman_sqlite()
  batting <- tbl(lahman_sqlite(), "Batting")
  batting
}

# Connect to a local postgres database with lahman database, if available
if (require("RPostgreSQL") && has_lahman("postgres")) {
  lahman_postgres()
  batting <- tbl(lahman_postgres(), "Batting")
}
}

