\name{manip}
\alias{arrange}
\alias{filter}
\alias{group}
\alias{manip}
\alias{mutate}
\alias{select}
\alias{summarise}
\title{Data manipulation functions.}
\usage{
  filter(.data, ...)

  summarise(.data, ...)

  mutate(.data, ...)

  arrange(.data, ...)

  select(.data, ...)

  group(.data, ...)
}
\arguments{
  \item{.data}{a data source}

  \item{...}{variables interpreted in the context of that
  data frame.}
}
\description{
  Data manipulation functions.
}

