\name{manip}
\alias{arrange}
\alias{filter}
\alias{group}
\alias{manip}
\alias{mutate}
\alias{select}
\alias{summarise}
\title{Data manipulation functions.}
\usage{
  filter(_data, ...)

  summarise(_data, ...)

  mutate(_data, ...)

  arrange(_data, ...)

  select(_data, ...)

  group(_data, ...)
}
\description{
  Data manipulation functions.
}

