% Generated by roxygen2 (4.0.1): do not edit by hand
\name{manip_dt}
\alias{arrange.data.table}
\alias{filter.data.table}
\alias{manip_dt}
\alias{mutate.data.table}
\alias{rename.data.table}
\alias{select.data.table}
\alias{summarise.data.table}
\title{Data manipulation for data tables.}
\usage{
\method{filter}{data.table}(.data, ..., .env = parent.frame())

\method{summarise}{data.table}(.data, ...)

\method{mutate}{data.table}(.data, ..., inplace = FALSE)

\method{arrange}{data.table}(.data, ...)

\method{select}{data.table}(.data, ...)

\method{rename}{data.table}(.data, ...)
}
\arguments{
\item{.data}{a data table}

\item{...}{variables interpreted in the context of \code{.data}}

\item{inplace}{if \code{FALSE} (the default) the data frame will be copied
prior to modification to avoid changes propagating via reference.}

\item{.env}{The environment in which to evaluate arguments not included
in the data. The default should suffice for ordinary usage.}
}
\description{
Data manipulation for data tables.
}
\examples{
if (require("data.table") && require("nycflights13")) {
# If you start with a data table, you end up with a data table
flights <- as.data.table(flights)
filter(flights, month == 1, day == 1, dest == "DFW")
head(select(flights, year:day))
summarise(flights, delay = mean(arr_delay, na.rm = TRUE), n = length(arr_delay))
head(mutate(flights, gained = arr_delay - dep_delay))
head(arrange(flights, dest, desc(arr_delay)))

# If you start with a tbl, you end up with a tbl
flights2 <- as.tbl(flights)
filter(flights2, month == 1, day == 1, dest == "DFW")
head(select(flights2, year:day))
summarise(flights2, delay = mean(arr_delay, na.rm = TRUE), n = length(arr_delay))
head(mutate(flights2, gained = arr_delay - dep_delay))
head(arrange(flights2, dest, desc(arr_delay)))
}
}

