% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{manip_grouped_dt}
\alias{arrange.grouped_dt}
\alias{filter.grouped_dt}
\alias{manip_grouped_dt}
\alias{mutate.grouped_dt}
\alias{rename.grouped_dt}
\alias{select_.grouped_dt}
\alias{summarise_.grouped_dt}
\title{Data manipulation for grouped data tables.}
\usage{
\method{filter}{grouped_dt}(.data, ...)

\method{summarise_}{grouped_dt}(.data, ..., .dots)

mutate.grouped_dt(.data, ..., inplace = FALSE)

arrange.grouped_dt(.data, ...)

\method{select_}{grouped_dt}(.data, args)

\method{rename}{grouped_dt}(.data, ...)
}
\arguments{
\item{.data}{a data table}

\item{...,args}{variables interpreted in the context of \code{.data}}

\item{inplace}{if \code{FALSE} (the default) the data frame will be copied
prior to modification to avoid changes propagating via reference.}
}
\description{
Data manipulation for grouped data tables.
}
\examples{
if (require("data.table") && require("nycflights13")) {
flights2 <- tbl_dt(flights)
by_dest <- group_by(flights2, dest)

filter(by_dest, arr_delay == max(arr_delay, na.rm = TRUE))
summarise(by_dest, arr = mean(arr_delay, na.rm = TRUE))

# Normalise arrival and departure delays by airport
scaled <- mutate(by_dest, arr_z = scale(arr_delay), dep_z = scale(dep_delay))
select(scaled, year:day, dest, arr_z:dep_z)

arrange(by_dest, desc(arr_delay))
select(by_dest, -(day:tailnum))

# All manip functions preserve grouping structure, except for summarise
# which removes a grouping level
by_day <- group_by(flights, year, month, day)
by_month <- summarise(by_day, delayed = sum(arr_delay > 0, na.rm = TRUE))
by_month
summarise(by_month, delayed = sum(delayed))

# You can also manually ungroup:
ungroup(by_day)
}
}

