% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mutate}
\alias{mutate}
\alias{transmute}
\title{Add new variables.}
\usage{
mutate(.data, ...)

transmute(.data, ...)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link{tbl_df}}, \code{\link{tbl_dt}} and \code{\link{tbl_sql}}.}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop
a variable.}
}
\value{
An object of the same class as \code{.data}.

  Data frame row names are silently dropped. To preserve, convert to an
  explicit variable.
}
\description{
Mutate adds new variables and preserves existing; transmute drops existing
variables.
}
\examples{
mutate(mtcars, displ_l = disp / 61.0237)
transmute(mtcars, displ_l = disp / 61.0237)

mutate(mtcars, cyl = NULL)
}
\seealso{
Other single.table.verbs: \code{\link{arrange}};
  \code{\link{filter}}; \code{\link{rename}},
  \code{\link{select}}; \code{\link{slice}};
  \code{\link{summarise}}, \code{\link{summarize}}
}

