% Generated by roxygen2 (4.0.0): do not edit by hand
\name{n}
\alias{n}
\title{The number of observations in the current group.}
\usage{
n()
}
\description{
This function is implemented special for each data source and can only
be used from within \code{\link{summarise}}, \code{\link{mutate}} and
\code{\link{filter}}
}
\examples{
data("hflights", package = "hflights")
carriers <- group_by(hflights, UniqueCarrier)
summarise(carriers, n())
mutate(carriers, n = n())
filter(carriers, n() == 79)
}

