% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{n}
\alias{n}
\title{The number of observations in the current group.}
\usage{
n()
}
\description{
This function is implemented specifically for each data source and can only
be used from within \code{\link[=summarise]{summarise()}}, \code{\link[=mutate]{mutate()}} and
\code{\link[=filter]{filter()}}.
}
\examples{
if (require("nycflights13")) {
carriers <- group_by(flights, carrier)
summarise(carriers, n())
mutate(carriers, n = n())
filter(carriers, n() < 100)
}
}
