% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_distinct}
\alias{n_distinct}
\title{Efficiently count the number of unique values in a vector.}
\usage{
n_distinct(x, na_rm = FALSE)
}
\arguments{
\item{x}{a vector of values}

\item{na_rm}{if \code{TRUE} missing values don't count}
}
\description{
This is a faster and more concise equivalent of \code{length(unique(x))}
}
\examples{
x <- sample(1:10, 1e5, rep = TRUE)
length(unique(x))
n_distinct(x)
}

