% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/each.R
\name{pick}
\alias{pick}
\title{Select variables from the current group}
\usage{
pick(...)
}
\arguments{
\item{...}{tidy selection of columns}
}
\description{
Select a tibble made of a tidy selection of columns
and the rows of the current group
}
\examples{
iris \%>\%
  group_by(Species) \%>\%
  summarise(sepal = head(pick(starts_with("Sepal")), 2L))

}
