% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query}
\alias{MonetDBQuery-class}
\alias{Query-class}
\alias{query}
\title{Create a mutable query object.}
\usage{
query(con, sql, .vars)
}
\arguments{
\item{con}{a \code{DBOConnection}}

\item{sql}{a string containing an sql query.}
}
\description{
A query object is mutable wrapper around a \code{DBIResult} that caches
expensive operations, and insulates the rest of dplyr from the vagaries of
DBI and the individual database implementation.
}
\keyword{internal}

