% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-values.R
\name{recode-and-replace-values}
\alias{recode-and-replace-values}
\alias{recode_values}
\alias{replace_values}
\title{Recode and replace values}
\usage{
recode_values(x, ..., from = NULL, to = NULL, default = NULL, ptype = NULL)

replace_values(x, ..., from = NULL, to = NULL)
}
\description{
TODO
}
\seealso{
\code{\link[=case_when]{case_when()}}
}
