% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataframe.R
\name{recycle_each}
\alias{recycle}
\alias{recycle_each}
\alias{recycle_whole}
\title{Recycle a Vector}
\usage{
recycle_each(x, times = (n/length(x)), n = NULL)

recycle(x, n = NULL)

recycle_whole(x, n = NULL)
}
\arguments{
\item{x}{An (atomic, one-dimensional) vector.}

\item{n}{The desired output length. Note that, when
  evaluated within the context of a \code{data_frame}
  call, \code{n} will automatically be filled with
  the resultant number of rows.}
}
\description{
Helper functions (primarily for \code{\link{data_frame}})
that enable explicit recycling for vectors.
}

