% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tibble.r, R/reexport-tidyselect.R,
%   R/utils-tidy-eval.R, R/utils.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{data_frame}
\alias{data_frame_}
\alias{as_data_frame}
\alias{lst}
\alias{lst_}
\alias{add_row}
\alias{type_sum}
\alias{glimpse}
\alias{frame_data}
\alias{tribble}
\alias{tibble}
\alias{as_tibble}
\alias{trunc_mat}
\alias{tbl_sum}
\alias{contains}
\alias{select_helpers}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{last_col}
\alias{quo}
\alias{quos}
\alias{quo_name}
\alias{ensym}
\alias{ensyms}
\alias{enexpr}
\alias{enexprs}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{quo}}, \code{\link[rlang]{quos}}, \code{\link[rlang]{quo_name}}, \code{\link[rlang]{ensym}}, \code{\link[rlang]{ensyms}}, \code{\link[rlang]{enexpr}}, \code{\link[rlang]{enexprs}}}

  \item{tibble}{\code{\link[tibble]{data_frame}}, \code{\link[tibble]{data_frame_}}, \code{\link[tibble]{as_data_frame}}, \code{\link[tibble]{lst}}, \code{\link[tibble]{lst_}}, \code{\link[tibble]{add_row}}, \code{\link[tibble]{type_sum}}, \code{\link[tibble]{glimpse}}, \code{\link[tibble]{frame_data}}, \code{\link[tibble]{tribble}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{as_tibble}}, \code{\link[tibble]{trunc_mat}}, \code{\link[tibble]{tbl_sum}}}

  \item{tidyselect}{\code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{last_col}}}
}}

