% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble-reexport.r, R/utils.r
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{add_row}
\alias{as_data_frame}
\alias{data_frame}
\alias{data_frame_}
\alias{frame_data}
\alias{glimpse}
\alias{lst}
\alias{lst_}
\alias{reexports}
\alias{tbl_df}
\alias{tibble}
\alias{trunc_mat}
\alias{type_sum}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{tbl_df}}, \code{\link[tibble]{data_frame}}, \code{\link[tibble]{data_frame_}}, \code{\link[tibble]{as_data_frame}}, \code{\link[tibble]{lst}}, \code{\link[tibble]{lst_}}, \code{\link[tibble]{add_row}}, \code{\link[tibble]{type_sum}}, \code{\link[tibble]{glimpse}}, \code{\link[tibble]{frame_data}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{trunc_mat}}}
}}
\keyword{internal}

