% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{reexports}
\alias{\%>\%}
\alias{reexports}
\title{Re-exported functions}
\description{
These functions have been imported from other packages, and reimported by this package so that you can use them without having to attach another package. See the original documentation, linked to below, for more details
}
\seealso{
\code{\link[magrittr]{\%>\%}}, 
}
\keyword{internal}

