% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{rows}
\alias{rows}
\title{List of rows of each group}
\usage{
rows(.data)
}
\arguments{
\item{.data}{a tibble}
}
\value{
a list of integer vectors, giving the indices of the rows of each group
}
\description{
List of rows of each group
}
\examples{
df <- tibble(x = c(1,1,2,2))
rows(df)
rows(group_by(df, x))
}
