% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{same_src}
\alias{same_src}
\title{Figure out if two sources are the same (or two tbl have the same source)}
\usage{
same_src(x, y)
}
\arguments{
\item{x,y}{src or tbls to test}
}
\value{
a logical flag
}
\description{
Figure out if two sources are the same (or two tbl have the same source)
}
\keyword{internal}

