% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tidyselect.R
\name{select_vars}
\alias{select_vars}
\alias{rename_vars}
\alias{select_var}
\title{Select variables}
\usage{
select_vars(vars, ..., include = chr(), exclude = chr())

rename_vars(vars, ..., strict = TRUE)

select_var(vars, var = -1)
}
\arguments{
\item{vars}{A character vector of existing column names.}

\item{...}{Expressions to compute.}

\item{include, exclude}{Character vector of column names to always
include/exclude.}

\item{strict}{If \code{TRUE}, will throw an error if you attempt to
rename a variable that doesn't exist.}

\item{var}{A variable specified as in the same argument of
\code{\link[tidyselect:vars_pull]{tidyselect::vars_pull()}}.}
}
\description{
These functions now live in the tidyselect package as
\code{\link[tidyselect:vars_select]{tidyselect::vars_select()}}, \code{\link[tidyselect:vars_rename]{tidyselect::vars_rename()}} and
\code{\link[tidyselect:vars_pull]{tidyselect::vars_pull()}}. These dplyr aliases will be deprecated
sometimes in the future.
}
