% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dbplyr.R
\name{sql}
\alias{sql}
\title{SQL escaping.}
\usage{
sql(...)
}
\arguments{
\item{...}{Character vectors that will be combined into a single SQL
expression.}
}
\description{
These functions are critical when writing functions that translate R
functions to sql functions. Typically a conversion function should escape
all its inputs and return an sql object.
}
