% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-build.R
\name{sql_build}
\alias{select_query}
\alias{sql_build}
\alias{sql_render}
\title{Build and render SQL from a sequence of lazy operations}
\usage{
sql_build(op, ...)

select_query(from, select = sql("*"), where = character(),
  group_by = character(), having = character(), order_by = character(),
  limit = NULL, offset = NULL)

sql_render(x, con = NULL, ...)
}
\arguments{
\item{op}{A sequence of lazy operations}

\item{...}{Other arguments passed on to the methods. Not currently used.}

\item{con}{A database connection. The default \code{NULL} uses a set of
rules that should be very similar to ANSI 92, and allows for testing
without an active database connection.}
}
\description{
\code{sql_build} creates a \code{select_query} S3 object, that is rendered
to a SQL string by \code{sql_render}. \code{sql_build} is particularly
well suited for testing.
}
\details{
\code{sql_build} is generic over the lazy operations, \link{lazy_ops}.
If a database does not follow regular ANSI 92 rules for creating a select
query, you can provide a methods for \code{sql_select} and
\code{sql_subquery}.
}
\keyword{internal}

