% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_dbi.R
\name{src_dbi}
\alias{src_dbi}
\alias{tbl.src_dbi}
\title{Create a source for a DBI backend.}
\usage{
src_dbi(con, auto_disconnect = FALSE)

\method{tbl}{src_dbi}(src, from, ...)
}
\arguments{
\item{con}{An object that inherits from \link[DBI:DBIConnection-class]{DBI::DBIConnection},
typically generated by \link[DBI:dbConnect]{DBI::dbConnect}}

\item{auto_disconnect}{Should the connection be automatically closed when
the src is deleted. This is useful for older DBI backends that don't
clean up themselves.}

\item{src}{A DBI src.}

\item{from}{Either a string giving the name of table in database, or
\code{\link[=sql]{sql()}} describing a derived table or compound join.}

\item{...}{Included for compatibility with the generic, but otherwise
ignored.}
}
\value{
An S3 object of class \code{src_dbi}, \code{src_sql}, \code{src}.
}
\description{
\code{src_dbi()} allows you to construct a dplyr src for an any DBI backends.
}
\examples{
if (requireNamespace("RSQLite", quietly = TRUE)) {

con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
src <- src_dbi(con)

copy_to(src, mtcars)

DBI::dbListTables(con)
src
}
}
