% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-src-local.r
\name{src_local}
\alias{src_local}
\alias{src_df}
\title{A local source}
\usage{
src_local(tbl, pkg = NULL, env = NULL)

src_df(pkg = NULL, env = NULL)
}
\arguments{
\item{tbl}{name of the function used to generate \code{tbl} objects}

\item{pkg, env}{Either the name of a package or an environment object in
which to look for objects.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
This function was deprecated since it existed to support a style of testing
dplyr backends that turned out not to be useful.
}
\keyword{internal}
