% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\name{src_tbls}
\alias{src_tbls}
\title{List all tbls provided by a source.}
\usage{
src_tbls(x, ...)
}
\arguments{
\item{x}{a data src.}

\item{...}{other arguments passed on to the individual methods.}
}
\description{
This is a generic method which individual src's will provide methods for.
Most methods will not be documented because it's usually pretty obvious what
possible results will be.
}
\keyword{internal}
