% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{summarise_each}
\alias{mutate_each}
\alias{mutate_each_}
\alias{summarise_each}
\alias{summarise_each_}
\alias{summarize_each}
\alias{summarize_each_}
\title{Deprecated _each() variants}
\usage{
summarise_each(tbl, funs, ...)

summarise_each_(tbl, funs, vars)

summarize_each(tbl, funs, ...)

summarize_each_(tbl, funs, vars)

mutate_each(tbl, funs, ...)

mutate_each_(tbl, funs, vars)
}
\arguments{
\item{tbl}{a tbl}

\item{funs}{List of function calls, generated by \code{\link{funs}}, or
a character vector of function names.}

\item{vars, ...}{Variables to include/exclude in mutate/summarise.
  You can use same specifications as in \code{\link{select}}. If missing,
  defaults to all non-grouping variables.

  For standard evaluation versions (ending in \code{_}) these can
  be either a list of expressions or a character vector.}
}
\description{
\code{mutate_each()} and \code{summarise_each()} are deprecated in
favour of a more featureful family of functions:
\code{\link{mutate_all}()}, \code{\link{mutate_at}()},
\code{\link{mutate_if}()}, \code{\link{summarise_all}()},
\code{\link{summarise_at}()} and \code{\link{summarise_if}()}.
}

