% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{summarise_each}
\alias{mutate_each}
\alias{mutate_each_}
\alias{mutate_each_q}
\alias{summarise_each}
\alias{summarise_each_}
\alias{summarise_each_q}
\alias{summarize_each}
\alias{summarize_each_}
\title{Summarise and mutate multiple columns.}
\usage{
summarise_each(tbl, funs, ...)

summarise_each_(tbl, funs, vars)

summarize_each(tbl, funs, ...)

summarize_each_(tbl, funs, vars)

mutate_each(tbl, funs, ...)

mutate_each_(tbl, funs, vars)
}
\arguments{
\item{tbl}{a tbl}

\item{funs}{List of function calls, generated by \code{\link{funs}}, or
a character vector of function names.}

\item{...}{For standard evaluation versions (ending in \code{_}) these can
be either a list of expressions or a character vector.}

\item{vars}{Variables to include/exclude in mutate/summarise.
You can use same specifications as in \code{\link{select}}. If missing,
defaults to all non-grouping variables.}
}
\description{
Apply one or more functions to one or more columns. Grouping variables
are always excluded from modification.
}
\examples{
# One function
by_species <- iris \%>\% group_by(Species)
by_species \%>\% summarise_each(funs(length))
by_species \%>\% summarise_each(funs(mean))
by_species \%>\% summarise_each(funs(mean), Petal.Width)
by_species \%>\% summarise_each(funs(mean), matches("Width"))

by_species \%>\% mutate_each(funs(half = . / 2))
by_species \%>\% mutate_each(funs(min_rank))

# Two functions
by_species \%>\% summarise_each(funs(min, max))
by_species \%>\% summarise_each(funs(min, max), Petal.Width, Sepal.Width)
by_species \%>\% summarise_each(funs(min, max), matches("Width"))

# Alternative function specification
iris \%>\% summarise_each(funs(ul = length(unique(.))))
by_species \%>\% summarise_each(funs(ul = length(unique(.))))

by_species \%>\% summarise_each(c("min", "max"))

# Alternative variable specification
summarise_each_(iris, funs(max), names(iris)[-5])
summarise_each_(iris, funs(max), list(quote(-Species)))
}

