% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-lazyeval.R
\name{summarise_each}
\alias{summarise_each}
\alias{summarise_each_}
\alias{mutate_each}
\alias{mutate_each_}
\alias{summarize_each}
\alias{summarize_each_}
\title{Summarise and mutate multiple columns.}
\usage{
summarise_each(tbl, funs, ...)

summarise_each_(tbl, funs, vars)

mutate_each(tbl, funs, ...)

mutate_each_(tbl, funs, vars)

summarize_each(tbl, funs, ...)

summarize_each_(tbl, funs, vars)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}

\code{mutate_each()} and \code{summarise_each()} are deprecated in favour of
the new \code{\link[=across]{across()}} function that works within \code{summarise()} and \code{mutate()}.
}
\keyword{internal}
