% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise-mutate.R
\name{summarise_each}
\alias{summarise_each}
\alias{summarise_each_}
\alias{mutate_each}
\alias{mutate_each_}
\alias{summarize_each}
\alias{summarize_each_}
\title{Summarise and mutate multiple columns.}
\usage{
summarise_each(tbl, funs, ...)

summarise_each_(tbl, funs, vars)

mutate_each(tbl, funs, ...)

mutate_each_(tbl, funs, vars)

summarize_each(tbl, funs, ...)

summarize_each_(tbl, funs, vars)
}
\description{
\code{mutate_each()} and \code{summarise_each()} are deprecated in favour of
a more featureful family of functions: \code{\link[=mutate_all]{mutate_all()}},
\code{\link[=mutate_at]{mutate_at()}}, \code{\link[=mutate_if]{mutate_if()}}, \code{\link[=summarise_all]{summarise_all()}}, \code{\link[=summarise_at]{summarise_at()}}
and \code{\link[=summarise_if]{summarise_if()}}.

The \verb{_each()} functions have two replacements depending on what
variables you want to apply \code{funs} to. To apply a function to all
variables, use \code{\link[=mutate_all]{mutate_all()}} or \code{\link[=summarise_all]{summarise_all()}}. To apply a
function to a selection of variables, use \code{\link[=mutate_at]{mutate_at()}} or
\code{\link[=summarise_at]{summarise_at()}}.

See the relevant section of \code{vignette("compatibility")} for more
information.
}
\details{
\Sexpr[results=rd, stage=render]{dplyr:::lifecycle("deprecated")}
}
\keyword{internal}
