\name{tbl}
\alias{as.tbl}
\alias{is.tbl}
\alias{tbl}
\title{Create a table from a data source}
\usage{
  tbl(src, ...)

  is.tbl(x)

  as.tbl(x, ...)
}
\arguments{
  \item{src}{A data source}

  \item{...}{Other arguments passed on to the individual
  methods}
}
\description{
  This is a generic method that dispatches based on the
  first argument.
}

