% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.r
\name{tbl}
\alias{tbl}
\alias{is.tbl}
\alias{as.tbl}
\title{Create a table from a data source}
\usage{
tbl(src, ...)

is.tbl(x)

as.tbl(x, ...)
}
\arguments{
\item{src}{A data source}

\item{...}{Other arguments passed on to the individual methods}

\item{x}{an object to coerce to a \code{tbl}}
}
\description{
This is a generic method that dispatches based on the first argument.
}
