% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.r
\name{tbl}
\alias{tbl}
\alias{is.tbl}
\title{Create a table from a data source}
\usage{
tbl(src, ...)

is.tbl(x)
}
\arguments{
\item{src}{A data source}

\item{...}{Other arguments passed on to the individual methods}

\item{x}{Any object}
}
\description{
This is a generic method that dispatches based on the first argument.
}
