% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-df.r
\name{tbl_df}
\alias{tbl_df}
\title{Create a data frame tbl.}
\usage{
tbl_df(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Deprecated: please use \code{\link[tibble:as_tibble]{tibble::as_tibble()}} instead.
}
\keyword{internal}
