% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-tibble.R
\name{tbl_df}
\alias{tbl_df}
\alias{as.tbl}
\title{Coerce to a tibble}
\usage{
tbl_df(data)

as.tbl(x, ...)
}
\arguments{
\item{data, x}{Object to coerce}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link[tibble:as_tibble]{tibble::as_tibble()}} instead.
}
\keyword{internal}
