% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-tibble.R
\name{tbl_df}
\alias{tbl_df}
\title{Create a data frame tbl.}
\usage{
tbl_df(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link[tibble:as_tibble]{tibble::as_tibble()}} instead.
}
\keyword{internal}
