% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.r
\name{tbl_vars}
\alias{tbl_vars}
\alias{tbl_nongroup_vars}
\title{List variables provided by a tbl.}
\usage{
tbl_vars(x)

tbl_nongroup_vars(x)
}
\arguments{
\item{x}{A tbl object}
}
\description{
\code{tbl_vars()} returns all variables while \code{tbl_nongroup_vars()}
returns only non-grouping variables. The \code{groups} attribute
of the object returned by \code{tbl_vars()} is a character vector of the
grouping columns.
}
\seealso{
\code{\link[=group_vars]{group_vars()}} for a function that returns grouping
variables.
}
\keyword{internal}
