% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{tidyeval-compat}
\alias{tidyeval-compat}
\alias{expr}
\alias{enquo}
\alias{enquos}
\alias{sym}
\alias{syms}
\alias{as_label}
\alias{quo}
\alias{quos}
\alias{quo_name}
\alias{ensym}
\alias{ensyms}
\alias{enexpr}
\alias{enexprs}
\title{Other tidy eval tools}
\description{
These tidy eval functions are no longer for normal usage, but are still
exported from dplyr for backward compatibility.
See \code{\link[=dplyr_data_masking]{?dplyr_data_masking}} and \code{vignette("programming")}
for the latest recommendations.
\itemize{
\item \link[rlang:nse-defuse]{expr()}
\item \link[rlang:nse-defuse]{enquo()}
\item \link[rlang:nse-defuse]{enquos()}
\item \link[rlang:sym]{sym()}
\item \link[rlang:sym]{syms()}
\item \link[rlang:as_label]{as_label()}
\item \link[rlang:nse-defuse]{quo()}
\item \link[rlang:nse-defuse]{quos()}
\item \link[rlang:quo_label]{quo_name()}
\item \link[rlang:nse-defuse]{ensym()}
\item \link[rlang:nse-defuse]{ensyms()}
\item \link[rlang:nse-defuse]{enexpr()}
\item \link[rlang:nse-defuse]{enexprs()}
}
}
\keyword{internal}
