% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{vars}
\alias{vars}
\title{Select variables}
\usage{
vars(...)
}
\arguments{
\item{...}{Variables to include/exclude in mutate/summarise. You
can use same specifications as in \code{\link[=select]{select()}}. If missing,
defaults to all non-grouping variables.

These arguments are automatically \link[rlang:quo]{quoted} and later
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting}. See
\code{vignette("programming")} for an introduction to these concepts.}
}
\description{
This helper is intended to provide equivalent semantics to
\code{\link[=select]{select()}}. It is used for instance in scoped summarising and
mutating verbs (\code{\link[=mutate_at]{mutate_at()}} and \code{\link[=summarise_at]{summarise_at()}}).
}
\details{
Note that verbs accepting a \code{vars()} specification also accept an
\link[rlang:is_integerish]{integerish} vector of positions or a
character vector of column names.
}
\seealso{
\code{\link[=funs]{funs()}}, \code{\link[=all_vars]{all_vars()}} and \code{\link[=any_vars]{any_vars()}} for other quoting
functions that you can use with scoped verbs.
}
