% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise.R
\name{vars}
\alias{vars}
\title{Select variables}
\usage{
vars(...)
}
\arguments{
\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Variables to include/exclude
in mutate/summarise. You can use same specifications as in \code{\link[=select]{select()}}.
If missing, defaults to all non-grouping variables.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}

\code{vars()} was only needed for the scoped verbs, which have been superseded
by the use of \code{\link[=across]{across()}} in an existing verb. See \code{vignette("colwise")} for
details.

This helper is intended to provide equivalent semantics to
\code{\link[=select]{select()}}. It is used for instance in scoped summarising and
mutating verbs (\code{\link[=mutate_at]{mutate_at()}} and \code{\link[=summarise_at]{summarise_at()}}).

Note that verbs accepting a \code{vars()} specification also accept a
numeric vector of positions or a character vector of column names.
}
\seealso{
\code{\link[=all_vars]{all_vars()}} and \code{\link[=any_vars]{any_vars()}} for other quoting
functions that you can use with scoped verbs.
}
